Module time::duration

The Duration struct and its associated impls.

Macros

duration_mul_div_int 🔒

Implement Mul (reflexively) and Div for Duration for various types.

try_from_secs 🔒

This is adapted from the std implementation, which uses mostly bit operations to ensure the highest precision:

Structs

Duration

A span of time with nanosecond precision.

Enums

Padding 🔒

By explicitly inserting this enum where padding is expected, the compiler is able to better perform niche value optimization.

Type Aliases

Nanoseconds 🔒

The type of the nanosecond field of Duration.