Enum time::Sign

#[repr(i8)]
pub enum Sign {
    Positive,
    Negative,
    Zero,
}

👎 Deprecated since 0.2.7:
The only use for this (obtaining the sign of a Duration) can be replaced with Duration::is_{positive|negative|zero}

Contains the sign of a value: positive, negative, or zero.

For ease of use, Sign implements Mul and Div on all signed numeric types. Signs can also be multiplied and divided by another Sign, which follows the same rules as real numbers.

Variants

Positive

👎 Deprecated since 0.2.7:
The only use for this (obtaining the sign of a Duration) can be replaced with Duration::is_{positive|negative|zero}

A positive value.

Negative

👎 Deprecated since 0.2.7:
The only use for this (obtaining the sign of a Duration) can be replaced with Duration::is_{positive|negative|zero}

A negative value.

Zero

👎 Deprecated since 0.2.7:
The only use for this (obtaining the sign of a Duration) can be replaced with Duration::is_{positive|negative|zero}

A value that is exactly zero.

Implementations

impl Sign

pub fn negate(self) -> Self

Return the opposite of the current sign.

assert_eq!(Sign::Positive.negate(), Sign::Negative);
assert_eq!(Sign::Negative.negate(), Sign::Positive);
assert_eq!(Sign::Zero.negate(), Sign::Zero);

pub const fn is_positive(self) -> bool

Is the sign positive?

assert!(Sign::Positive.is_positive());
assert!(!Sign::Negative.is_positive());
assert!(!Sign::Zero.is_positive());

pub const fn is_negative(self) -> bool

Is the sign negative?

assert!(!Sign::Positive.is_negative());
assert!(Sign::Negative.is_negative());
assert!(!Sign::Zero.is_negative());

pub const fn is_zero(self) -> bool

Is the value exactly zero?

assert!(!Sign::Positive.is_zero());
assert!(!Sign::Negative.is_zero());
assert!(Sign::Zero.is_zero());