time

Module error

Various error types returned by methods in the time crate.

Modules

component_range 🔒

Component range error

conversion_range 🔒

Conversion range error

different_variant 🔒

Different variant error

format 🔒

Error formatting a struct

indeterminate_offset 🔒

Indeterminate offset

invalid_format_description 🔒

Invalid format description

invalid_variant 🔒

Invalid variant error

parse 🔒

Error that occurred at some stage of parsing

parse_from_description 🔒

Error parsing an input into a Parsed struct

try_from_parsed 🔒

Error converting a Parsed struct to another type

Structs

ComponentRange

An error type indicating that a component provided to a method was out of range, causing a failure.

ConversionRange

An error type indicating that a conversion failed because the target type could not store the initial value.

DifferentVariant

An error type indicating that a TryFrom call failed because the original value was of a different variant.

IndeterminateOffset

The system’s UTC offset could not be determined at the given datetime.

InvalidVariant

An error type indicating that a FromStr call failed because the value was not a valid variant.

Enums

Error

A unified error type for anything returned by a method in the time crate.

Format

An error occurred when formatting.

InvalidFormatDescription

The format description provided was not valid.

Parse

An error that occurred at some stage of parsing.

ParseFromDescription

An error that occurred while parsing the input into a Parsed struct.

TryFromParsed

An error that occurred when converting a Parsed to another type.