time::error

Enum Format

#[non_exhaustive]
pub enum Format {
    InsufficientTypeInformation,
    InvalidComponent(&'static str),
    ComponentRange(Box<ComponentRange>),
    StdIo(Error),
}

Available on crate feature formatting only.

An error occurred when formatting.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

InsufficientTypeInformation

The type being formatted does not contain sufficient information to format a component.

InvalidComponent(&'static str)

The component named has a value that cannot be formatted into the requested format.

This variant is only returned when using well-known formats.

ComponentRange(Box<ComponentRange>)

A component provided was out of range.

StdIo(Error)

A value of std::io::Error was returned internally.