Enum time::error::InvalidFormatDescription

#[non_exhaustive]
pub enum InvalidFormatDescription {
    UnclosedOpeningBracket {
        index: usize,
    },
    InvalidComponentName {
        name: String,
        index: usize,
    },
    InvalidModifier {
        value: String,
        index: usize,
    },
    MissingComponentName {
        index: usize,
    },
    MissingRequiredModifier {
        name: &'static str,
        index: usize,
    },
    Expected {
        what: &'static str,
        index: usize,
    },
    NotSupported {
        what: &'static str,
        context: &'static str,
        index: usize,
    },
}

The format description provided was not valid.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

UnclosedOpeningBracket

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

index: usize

The zero-based index of the opening bracket.

There was a bracket pair that was opened but not closed.

InvalidComponentName

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

name: String

The name of the invalid component name.

index: usize

The zero-based index the component name starts at.

A component name is not valid.

InvalidModifier

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

value: String

The value of the invalid modifier.

index: usize

The zero-based index the modifier starts at.

A modifier is not valid.

MissingComponentName

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

index: usize

The zero-based index where the component name should start.

A component name is missing.

MissingRequiredModifier

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

name: &'static str

The name of the modifier that is missing.

index: usize

The zero-based index of the component.

A required modifier is missing.

Expected

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

what: &'static str

What was expected to be present, but wasn’t.

index: usize

The zero-based index the item was expected to be found at.

Something was expected, but not found.

NotSupported

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

what: &'static str

The behavior that is not supported.

context: &'static str

The context in which the behavior is not supported.

index: usize

The zero-based index the error occurred at.

Certain behavior is not supported in the given context.