Enum time::error::Parse

#[non_exhaustive]pub enum Parse {
    InvalidNanosecond,
    InvalidSecond,
    InvalidMinute,
    InvalidHour,
    InvalidAmPm,
    InvalidMonth,
    InvalidYear,
    InvalidWeek,
    InvalidDayOfWeek,
    InvalidDayOfMonth,
    InvalidDayOfYear,
    InvalidOffset,
    MissingFormatSpecifier,
    InvalidFormatSpecifier(char),
    UnexpectedCharacter {
        expected: char,
        actual: char,
    },
    UnexpectedEndOfString,
    InsufficientInformation,
    ComponentOutOfRange(Box<ComponentRange>),
}

An error occurred while parsing.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

InvalidNanosecond

The nanosecond present was not valid.

InvalidSecond

The second present was not valid.

InvalidMinute

The minute present was not valid.

InvalidHour

The hour present was not valid.

InvalidAmPm

The AM/PM was not valid.

InvalidMonth

The month present was not valid.

InvalidYear

The year present was not valid.

InvalidWeek

The week present was not valid.

InvalidDayOfWeek

The day of week present was not valid.

InvalidDayOfMonth

The day of month present was not valid.

InvalidDayOfYear

The day of year present was not valid.

InvalidOffset

The UTC offset present was not valid.

MissingFormatSpecifier

There was no character following a %.

InvalidFormatSpecifier(char)

The character following % is not valid.

UnexpectedCharacter

A character literal was expected to be present but was not.

Fields of UnexpectedCharacter

expected: char

The character that was expected to be present.

actual: char

The character that was present in the string.

UnexpectedEndOfString

The string ended, but there should be more content.

InsufficientInformation

There was not enough information provided to create the requested type.

ComponentOutOfRange(Box<ComponentRange>)

A component was out of range.