Struct time::error::ComponentRange

pub struct ComponentRange { /* private fields */ }

An error type indicating that a component provided to a method was out of range, causing a failure.

Implementations

impl ComponentRange

pub const fn name(self) -> &'static str

Obtain the name of the component whose value was out of range.

pub const fn is_conditional(self) -> bool

Whether the value’s permitted range is conditional, i.e. whether an input with this value could have succeeded if the values of other components were different.