time::ext::digit_count

Trait DigitCount

pub(crate) trait DigitCount {
    // Required method
    fn num_digits(self) -> u8;
}

A trait that indicates the formatted width of the value can be determined.

Note that this should not be implemented for any signed integers. This forces the caller to write the sign if desired.

Required Methods

fn num_digits(self) -> u8

The number of digits in the stringified value.