Function time::now

pub fn now<T>() -> T::Unitwhere
    T: ClockType,

Returns the current time.

Monotonic clocks return an Instant whereas wall time clocks return a Duration signifying the time since 12:00am January 1st 1970 (i.e. Unix time).

This function must not be called prior to registering a clock source of the specified type using register_clock_source.