Function time::register_early_sleeper

pub fn register_early_sleeper<T>(frequency: Frequency) -> boolwhere
    T: EarlySleeper,

Register the clock source that can be used to sleep when interrupts are disabled.

The current early sleeper will only be overwritten by T if frequency is larger than the frequency of the current early sleeper.

Returns whether the early sleeper was overwritten.