time::format_description

Enum Component

#[non_exhaustive]
pub enum Component {

    Day(Day),
    MonthShort(MonthShort),
    MonthLong(MonthLong),
    MonthNumerical(MonthNumerical),
    Ordinal(Ordinal),
    WeekdayShort(WeekdayShort),
    WeekdayLong(WeekdayLong),
    WeekdaySunday(WeekdaySunday),
    WeekdayMonday(WeekdayMonday),
    WeekNumberIso(WeekNumberIso),
    WeekNumberSunday(WeekNumberSunday),
    WeekNumberMonday(WeekNumberMonday),
    CalendarYearFullExtendedRange(CalendarYearFullExtendedRange),
    CalendarYearFullStandardRange(CalendarYearFullStandardRange),
    IsoYearFullExtendedRange(IsoYearFullExtendedRange),
    IsoYearFullStandardRange(IsoYearFullStandardRange),
    CalendarYearCenturyExtendedRange(CalendarYearCenturyExtendedRange),
    CalendarYearCenturyStandardRange(CalendarYearCenturyStandardRange),
    IsoYearCenturyExtendedRange(IsoYearCenturyExtendedRange),
    IsoYearCenturyStandardRange(IsoYearCenturyStandardRange),
    CalendarYearLastTwo(CalendarYearLastTwo),
    IsoYearLastTwo(IsoYearLastTwo),
    Hour12(Hour12),
    Hour24(Hour24),
    Minute(Minute),
    Period(Period),
    Second(Second),
    Subsecond(Subsecond),
    OffsetHour(OffsetHour),
    OffsetMinute(OffsetMinute),
    OffsetSecond(OffsetSecond),
    Ignore(Ignore),
    UnixTimestampSecond(UnixTimestampSecond),
    UnixTimestampMillisecond(UnixTimestampMillisecond),
    UnixTimestampMicrosecond(UnixTimestampMicrosecond),
    UnixTimestampNanosecond(UnixTimestampNanosecond),
    End(End),
    Month(Month),
    Weekday(Weekday),
    WeekNumber(WeekNumber),
    Hour(Hour),
    UnixTimestamp(UnixTimestamp),
    Year(Year),

}

A component of a larger format description.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Day(Day)

Day of the month.

MonthShort(MonthShort)

Month of the year in the abbreviated form (e.g. “Jan”).

MonthLong(MonthLong)

Month of the year in the full form (e.g. “January”).

MonthNumerical(MonthNumerical)

Month of the year in the numerical form (e.g. “1” for January).

Ordinal(Ordinal)

Ordinal day of the year.

WeekdayShort(WeekdayShort)

Weekday in the abbreviated form (e.g. “Mon”).

WeekdayLong(WeekdayLong)

Weekday in the full form (e.g. “Monday”).

WeekdaySunday(WeekdaySunday)

Weekday number where Sunday is either 0 or 1 depending on the modifier.

WeekdayMonday(WeekdayMonday)

Weekday number where Monday is either 0 or 1 depending on the modifier.

WeekNumberIso(WeekNumberIso)

Week number of the year, where week 1 starts is the week beginning on Monday that contains January 4.

WeekNumberSunday(WeekNumberSunday)

Week number of the year, where week 1 starts on the first Sunday of the calendar year.

WeekNumberMonday(WeekNumberMonday)

Week number of the year, where week 1 starts on the first Monday of the calendar year.

CalendarYearFullExtendedRange(CalendarYearFullExtendedRange)

The calendar year. Supports the extended range.

CalendarYearFullStandardRange(CalendarYearFullStandardRange)

The calendar year. Does not support the extended range.

IsoYearFullExtendedRange(IsoYearFullExtendedRange)

The ISO week-based year. Supports the extended range.

IsoYearFullStandardRange(IsoYearFullStandardRange)

The ISO week-based year. Does not support the extended range.

CalendarYearCenturyExtendedRange(CalendarYearCenturyExtendedRange)

The century of the calendar year. Supports the extended range.

CalendarYearCenturyStandardRange(CalendarYearCenturyStandardRange)

The century of the calendar year. Does not support the extended range.

IsoYearCenturyExtendedRange(IsoYearCenturyExtendedRange)

The century of the ISO week-based year. Supports the extended range.

IsoYearCenturyStandardRange(IsoYearCenturyStandardRange)

The century of the ISO week-based year. Does not support the extended range.

CalendarYearLastTwo(CalendarYearLastTwo)

The last two digits of the calendar year.

IsoYearLastTwo(IsoYearLastTwo)

The last two digits of the ISO week-based year.

Hour12(Hour12)

Hour of the day using the 12-hour clock.

Hour24(Hour24)

Hour of the day using the 24-hour clock.

Minute(Minute)

Minute within the hour.

Period(Period)

AM/PM part of the time.

Second(Second)

Second within the minute.

Subsecond(Subsecond)

Subsecond within the second.

OffsetHour(OffsetHour)

Hour of the UTC offset.

OffsetMinute(OffsetMinute)

Minute within the hour of the UTC offset.

OffsetSecond(OffsetSecond)

Second within the minute of the UTC offset.

Ignore(Ignore)

A number of bytes to ignore when parsing. This has no effect on formatting.

UnixTimestampSecond(UnixTimestampSecond)

A Unix timestamp in seconds.

UnixTimestampMillisecond(UnixTimestampMillisecond)

A Unix timestamp in milliseconds.

UnixTimestampMicrosecond(UnixTimestampMicrosecond)

A Unix timestamp in microseconds.

UnixTimestampNanosecond(UnixTimestampNanosecond)

A Unix timestamp in nanoseconds.

End(End)

The end of input. Parsing this component will fail if there is any input remaining. This component neither affects formatting nor consumes any input when parsing.

Month(Month)

👎Deprecated since 0.3.48:
use MonthShort, MonthLong, or MonthNumeric instead

Month of the year.

Weekday(Weekday)

👎Deprecated since 0.3.48:
use WeekdayShort, WeekdayLong, or WeekdaySunday, or WeekdayMonday instead

Day of the week.

WeekNumber(WeekNumber)

👎Deprecated since 0.3.48:
use WeekNumberIso, WeekNumberSunday, or WeekNumberMonday instead

Week within the year.

Hour(Hour)

👎Deprecated since 0.3.48:
use Hour12 or Hour24 instead

Hour of the day.

UnixTimestamp(UnixTimestamp)

👎Deprecated since 0.3.48:
use UnixTimestampSeconds, UnixTimestampMilliseconds, UnixTimestampMicroseconds, or UnixTimestampNanoseconds instead

A Unix timestamp.

Year(Year)

👎Deprecated since 0.3.48:
use one of the various Year* components instead

Year of the date.