Enum time::format_description::FormatItem

#[non_exhaustive]
pub enum FormatItem<'a> {
    Literal(&'a [u8]),
    Component(Component),
    Compound(&'a [Self]),
    Optional(&'a Self),
    First(&'a [Self]),
}

A complete description of how to format and parse a type.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Literal(&'a [u8])

Bytes that are formatted as-is.

Note: If you call the format method that returns a String, these bytes will be passed through String::from_utf8_lossy.

Tuple Fields of Literal

0: &'a [u8]

Component(Component)

A minimal representation of a single non-literal item.

Tuple Fields of Component

0: Component

Compound(&'a [Self])

A series of literals or components that collectively form a partial or complete description.

Tuple Fields of Compound

0: &'a [Self]

Optional(&'a Self)

A FormatItem that may or may not be present when parsing. If parsing fails, there will be no effect on the resulting struct.

This variant has no effect on formatting, as the value is guaranteed to be present.

Tuple Fields of Optional

0: &'a Self

First(&'a [Self])

A series of FormatItems where, when parsing, the first successful parse is used. When formatting, the first element of the slice is used. An empty slice is a no-op when formatting or parsing.

Tuple Fields of First

0: &'a [Self]