Struct time::format_description::modifier::Minute

#[non_exhaustive]
pub struct Minute {
    pub padding: Padding,
}

Minute within the hour.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

padding: Padding

The padding to obtain the minimum width.

Implementations

impl Minute

pub const fn default() -> Self

Creates a modifier that indicates the value is padded with zeroes.

This function exists since [Default::default()] cannot be used in a const context. It may be removed once that becomes possible. As the [Default] trait is in the prelude, removing this function in the future will not cause any resolution failures for the overwhelming majority of users; only users who use #![no_implicit_prelude] will be affected. As such it will not be considered a breaking change.