time::format_description

Module parse

Parser for format descriptions.

Re-exports

pub use self::strftime::parse_strftime_borrowed;

pub use self::strftime::parse_strftime_owned;

Modules

format_item 🔒

Typed, validated representation of a parsed format description.

lexer_ast 🔒

Lexer for parsing format descriptions.

sealed 🔒

strftime 🔒

Macros

assert_version 🔒

version 🔒

Structs

Error 🔒

A complete error description.

ErrorInner 🔒

The internal error type.

Location 🔒

A location within a string.

Span 🔒

A start and end point within a string.

Spanned 🔒

A value with an associated Span.

Unused 🔒

A value that may be used in the future, but currently is not.

WithLocation 🔒

A value with an associated Location.

Traits

OptionExt 🔒

SpannedValue 🔒

Helper trait to attach a Span to a value.

WithLocationValue 🔒

Helper trait to attach a Location to a value.

Functions

parseDeprecated

Parse a sequence of items from the format description.

parse_borrowed

Parse a sequence of items from the format description.

parse_owned

Parse a sequence of items from the format description.

unused 🔒

Indicate that a value is currently unused.