Module time::format_description::parse::ast

AST for parsing format descriptions.

Structs

Modifier 🔒

A modifier for a component.

NestedFormatDescription 🔒

A format description that is nested within another format description.

Enums

Item 🔒

One part of a complete format description.

Functions

parse 🔒

Parse the provided tokens into an AST.

parse_component 🔒

Parse a component. This assumes that the opening bracket has already been consumed.

parse_inner 🔒

Parse the provided tokens into an AST. The const generic indicates whether the resulting Item will be used directly or as part of a NestedFormatDescription.

parse_nested 🔒

Parse a nested format description. The location provided is the the most recent one consumed.