Enum time::format_description::parse::ast::Item

pub(super) enum Item<'a> {
    Literal(Spanned<&'a [u8]>),
    EscapedBracket {
        _first: Unused<Location>,
        _second: Unused<Location>,
    },
    Component {
        _opening_bracket: Unused<Location>,
        _leading_whitespace: Unused<Option<Spanned<&'a [u8]>>>,
        name: Spanned<&'a [u8]>,
        modifiers: Box<[Modifier<'a>]>,
        _trailing_whitespace: Unused<Option<Spanned<&'a [u8]>>>,
        _closing_bracket: Unused<Location>,
    },
    Optional {
        opening_bracket: Location,
        _leading_whitespace: Unused<Option<Spanned<&'a [u8]>>>,
        _optional_kw: Unused<Spanned<&'a [u8]>>,
        _whitespace: Unused<Spanned<&'a [u8]>>,
        nested_format_description: NestedFormatDescription<'a>,
        closing_bracket: Location,
    },
    First {
        opening_bracket: Location,
        _leading_whitespace: Unused<Option<Spanned<&'a [u8]>>>,
        _first_kw: Unused<Spanned<&'a [u8]>>,
        _whitespace: Unused<Spanned<&'a [u8]>>,
        nested_format_descriptions: Box<[NestedFormatDescription<'a>]>,
        closing_bracket: Location,
    },
}

One part of a complete format description.

Variants

Literal(Spanned<&'a [u8]>)

A literal string, formatted and parsed as-is.

This should never be present inside a nested format description.

EscapedBracket

Fields

_first: Unused<Location>

The first bracket.

_second: Unused<Location>

The second bracket.

A sequence of brackets. The first acts as the escape character.

This should never be present if the lexer has BACKSLASH_ESCAPE set to true.

Component

Fields

_opening_bracket: Unused<Location>

Where the opening bracket was in the format string.

_leading_whitespace: Unused<Option<Spanned<&'a [u8]>>>

Whitespace between the opening bracket and name.

name: Spanned<&'a [u8]>

The name of the component.

modifiers: Box<[Modifier<'a>]>

The modifiers for the component.

_trailing_whitespace: Unused<Option<Spanned<&'a [u8]>>>

Whitespace between the modifiers and closing bracket.

_closing_bracket: Unused<Location>

Where the closing bracket was in the format string.

Part of a type, along with its modifiers.

Optional

Fields

opening_bracket: Location

Where the opening bracket was in the format string.

_leading_whitespace: Unused<Option<Spanned<&'a [u8]>>>

Whitespace between the opening bracket and “optional”.

_optional_kw: Unused<Spanned<&'a [u8]>>

The “optional” keyword.

_whitespace: Unused<Spanned<&'a [u8]>>

Whitespace between the “optional” keyword and the opening bracket.

nested_format_description: NestedFormatDescription<'a>

The items within the optional sequence.

closing_bracket: Location

Where the closing bracket was in the format string.

An optional sequence of items.

First

Fields

opening_bracket: Location

Where the opening bracket was in the format string.

_leading_whitespace: Unused<Option<Spanned<&'a [u8]>>>

Whitespace between the opening bracket and “first”.

_first_kw: Unused<Spanned<&'a [u8]>>

The “first” keyword.

_whitespace: Unused<Spanned<&'a [u8]>>

Whitespace between the “first” keyword and the opening bracket.

nested_format_descriptions: Box<[NestedFormatDescription<'a>]>

The sequences of items to try.

closing_bracket: Location

Where the closing bracket was in the format string.

The first matching parse of a sequence of items.