Struct time::format_description::well_known::iso8601::Config

pub struct Config { /* private fields */ }

Configuration for Iso8601.

Implementations

impl Config

pub const fn encode(&self) -> EncodedConfig

Encode the configuration, permitting it to be used as a const parameter of Iso8601.

The value returned by this method must only be used as a const parameter to Iso8601. Any other usage is unspecified behavior.

impl Config

pub const DEFAULT: Self = _

A configuration for the Iso8601 format.

The following is the default behavior:

The configuration can be used for both formatting and parsing.

The date, time, and UTC offset are all formatted.

Separators (such as - and :) are included.

The year contains four digits, such that the year must be between 0 and 9999.

The date uses the calendar format.

The time has precision to the second and nine decimal digits.

The UTC offset has precision to the minute.

If you need different behavior, use the setter methods on this struct.

pub const fn set_formatted_components(
self,
formatted_components: FormattedComponents
) -> Self

Set whether the format the date, time, and/or UTC offset.

pub const fn set_use_separators(self, use_separators: bool) -> Self

Set whether the format contains separators (such as - or :).

pub const fn set_year_is_six_digits(self, year_is_six_digits: bool) -> Self

Set whether the year is six digits.

pub const fn set_date_kind(self, date_kind: DateKind) -> Self

Set the format used for the date.

pub const fn set_time_precision(self, time_precision: TimePrecision) -> Self

Set the precision and number of decimal digits present for the time.

pub const fn set_offset_precision(
self,
offset_precision: OffsetPrecision
) -> Self

Set the precision for the UTC offset.