Struct time::format_description::well_known::iso8601::Iso8601

pub struct Iso8601<const CONFIG: EncodedConfig = { Config::DEFAULT.encode() }>;

The format described in ISO 8601.

This implementation is of ISO 8601-1:2019. It may not be compatible with other versions.

The const parameter CONFIG must be a value that was returned by Config::encode. Passing any other value is unspecified behavior.

Example: 1997-11-21T09:55:06.000000000-06:00

Examples

assert_eq!(
    datetime!(1997-11-12 9:55:06 -6:00).format(&Iso8601::DEFAULT)?,
    "1997-11-12T09:55:06.000000000-06:00"
);
Run

Implementations

impl Iso8601<DEFAULT>

pub const DEFAULT: Self = Self

An Iso8601 with the default configuration.

The following is the default behavior:

The configuration can be used for both formatting and parsing.

The date, time, and UTC offset are all formatted.

Separators (such as - and :) are included.

The year contains four digits, such that the year must be between 0 and 9999.

The date uses the calendar format.

The time has precision to the second and nine decimal digits.

The UTC offset has precision to the minute.

If you need different behavior, use another associated constant. For full customization, use Config::DEFAULT and Config’s methods to create a custom configuration.

impl Iso8601<PARSING>

pub const PARSING: Self = Self

An Iso8601 that can only be used for parsing. Using this to format a value is unspecified behavior.

impl Iso8601<DATE>

pub const DATE: Self = Self

An Iso8601 that handles only the date, but is otherwise the same as Config::DEFAULT.

impl Iso8601<TIME>

pub const TIME: Self = Self

An Iso8601 that handles only the time, but is otherwise the same as Config::DEFAULT.

impl Iso8601<OFFSET>

pub const OFFSET: Self = Self

An Iso8601 that handles only the UTC offset, but is otherwise the same as Config::DEFAULT.

impl Iso8601<DATE_TIME>

pub const DATE_TIME: Self = Self

An Iso8601 that handles the date and time, but is otherwise the same as Config::DEFAULT.

impl Iso8601<DATE_TIME_OFFSET>

pub const DATE_TIME_OFFSET: Self = Self

An Iso8601 that handles the date, time, and UTC offset. This is the same as Config::DEFAULT.

impl Iso8601<TIME_OFFSET>

pub const TIME_OFFSET: Self = Self

An Iso8601 that handles the time and UTC offset, but is otherwise the same as Config::DEFAULT.