time

Module formatting

Formatting for various types.

Modules

component_provider 🔒

formattable 🔒

A trait that can be used to format an item from its components.

iso8601 🔒

Helpers for implementing formatting for ISO 8601.

metadata 🔒

Constants

MONTH_NAMES 🔒

WEEKDAY_NAMES 🔒

Traits

Formattable

A type that describes a format.

Functions

f64_10_pow_x 🔒

Helper function to obtain 10^x, guaranteeing determinism for x ≤ 9. For these cases, the function optimizes to a lookup table. For x ≥ 10, it falls back to 10_f64.powi(x). The only situation where this would occur is if the user explicitly requests such precision when configuring the ISO 8601 well known format. All other possibilities max out at nine digits.

fmt_calendar_year_century_extended_range 🔒

fmt_calendar_year_century_standard_range 🔒

fmt_calendar_year_full_extended_range 🔒

fmt_calendar_year_full_standard_range 🔒

fmt_calendar_year_last_two 🔒

fmt_day 🔒

Format the day into the designated output.

fmt_hour_12 🔒

Format the hour into the designated output using the 12-hour clock.

fmt_hour_24 🔒

Format the hour into the designated output using the 24-hour clock.

fmt_iso_year_century_extended_range 🔒

fmt_iso_year_century_standard_range 🔒

fmt_iso_year_full_extended_range 🔒

fmt_iso_year_full_standard_range 🔒

fmt_iso_year_last_two 🔒

fmt_minute 🔒

Format the minute into the designated output.

fmt_month_long 🔒

Format the month into the designated output using the full name.

fmt_month_numerical 🔒

Format the month into the designated output as a number from 1-12.

fmt_month_short 🔒

Format the month into the designated output using the abbreviated name.

fmt_offset_hour 🔒

Format the offset hour into the designated output.

fmt_offset_minute 🔒

Format the offset minute into the designated output.

fmt_offset_second 🔒

Format the offset second into the designated output.

fmt_ordinal 🔒

Format the ordinal into the designated output.

fmt_period 🔒

Format the period into the designated output.

fmt_second 🔒

Format the second into the designated output.

fmt_sign 🔒

fmt_subsecond 🔒

Format the subsecond into the designated output.

fmt_unix_timestamp_microsecond 🔒

Format the Unix timestamp (in microseconds) into the designated output.

fmt_unix_timestamp_millisecond 🔒

Format the Unix timestamp (in milliseconds) into the designated output.

fmt_unix_timestamp_nanosecond 🔒

Format the Unix timestamp (in nanoseconds) into the designated output.

fmt_unix_timestamp_second 🔒

Format the Unix timestamp (in seconds) into the designated output.

fmt_week_number_iso 🔒

fmt_week_number_monday 🔒

fmt_week_number_sunday 🔒

fmt_weekday_long 🔒

Format the weekday into the designated output using the full name.

fmt_weekday_monday 🔒

Format the weekday into the designated output as a number from either 0-6 or 1-7 (depending on the modifier), where Monday is either 0 or 1.

fmt_weekday_short 🔒

Format the weekday into the designated output using the abbreviated name.

fmt_weekday_sunday 🔒

Format the weekday into the designated output as a number from either 0-6 or 1-7 (depending on the modifier), where Sunday is either 0 or 1.

format_five_digits_pad_zero 🔒

Format a five digit number that is padded with zeroes.

format_float 🔒

Write the floating point number to the output, returning the number of bytes written.

format_four_digits 🔒

Format a four digit number with the specified padding.

format_four_digits_pad_zero 🔒

Format a four digit number that is padded with zeroes.

format_int_padded 🔒

Write an integer with zeros as trailing padding if necessary to reach the requested width.

format_single_digit 🔒

Format a single digit.

format_six_digits_pad_zero 🔒

Format a six digit number that is padded with zeroes.

format_three_digits 🔒

Format a three digit number with the specified padding.

format_two_digits 🔒

Format a two digit number with the specified padding.

format_u64_pad_none 🔒

Format a number with no padding.

format_u128_pad_none 🔒

Format a number with no padding.

write 🔒

Write the string to the output, returning the number of bytes written.

write_bytes 🔒

Write all bytes to the output, returning the number of bytes written.

write_if 🔒

If pred is true, write the string to the output, returning the number of bytes written.

write_if_else 🔒

If pred is true, write true_str to the output. Otherwise, write false_str.

write_many 🔒

Write all strings to the output (in order), returning the total number of bytes written.

Type Aliases

Day 🔒

ExtendedCentury 🔒

IsoWeekNumber 🔒

LastTwo 🔒

MondayBasedWeek 🔒

OptionDay 🔒

OptionIsoWeekNumber 🔒

OptionYear 🔒

Ordinal 🔒

StandardCentury 🔒

StandardYear 🔒

SundayBasedWeek 🔒

Year 🔒