time::formatting::component_provider

Trait ComponentProvider

pub(crate) trait ComponentProvider {
    type State: Default;

    const SUPPLIES_DATE: bool = false;
    const SUPPLIES_TIME: bool = false;
    const SUPPLIES_OFFSET: bool = false;
    const SUPPLIES_TIMESTAMP: bool = false;

    // Provided methods
    fn day(&self, state: &mut Self::State) -> ru8<1, 31> { ... }
    fn month(&self, state: &mut Self::State) -> Month { ... }
    fn ordinal(&self, state: &mut Self::State) -> ru16<1, 366> { ... }
    fn weekday(&self, state: &mut Self::State) -> Weekday { ... }
    fn iso_week_number(&self, state: &mut Self::State) -> ru8<1, 53> { ... }
    fn monday_based_week(&self, state: &mut Self::State) -> ru8<0, 53> { ... }
    fn sunday_based_week(&self, state: &mut Self::State) -> ru8<0, 53> { ... }
    fn calendar_year(&self, state: &mut Self::State) -> ri32<-999_999, 999_999> { ... }
    fn iso_year(&self, state: &mut Self::State) -> ri32<-999_999, 999_999> { ... }
    fn hour(&self, state: &mut Self::State) -> ru8<0, { _ }> { ... }
    fn minute(&self, state: &mut Self::State) -> ru8<0, { _ }> { ... }
    fn period(&self, state: &mut Self::State) -> Period { ... }
    fn second(&self, state: &mut Self::State) -> ru8<0, { _ }> { ... }
    fn nanosecond(&self, state: &mut Self::State) -> ru32<0, { _ }> { ... }
    fn offset_is_negative(&self, state: &mut Self::State) -> bool { ... }
    fn offset_is_utc(&self, state: &mut Self::State) -> bool { ... }
    fn offset_hour(&self, state: &mut Self::State) -> ri8<-25, 25> { ... }
    fn offset_minute(&self, state: &mut Self::State) -> ri8<{ _ }, { _ }> { ... }
    fn offset_second(&self, state: &mut Self::State) -> ri8<{ _ }, { _ }> { ... }
    fn unix_timestamp_seconds(&self, state: &mut Self::State) -> i64 { ... }
    fn unix_timestamp_milliseconds(&self, state: &mut Self::State) -> i64 { ... }
    fn unix_timestamp_microseconds(&self, state: &mut Self::State) -> i128 { ... }
    fn unix_timestamp_nanoseconds(&self, state: &mut Self::State) -> i128 { ... }

}

A type with the ability to provide date, time, offset, and/or timestamp components on demand.

Note that while all methods have a default body, implementations are expected to override the body for all components that they provide. The default implementation exists solely for convenience, avoiding the need to specify unprovided components.

Provided Associated Constants

const SUPPLIES_DATE: bool = false

Whether the type can provide date components, indicating that date-related methods can be called.

const SUPPLIES_TIME: bool = false

Whether the type can provide time components, indicating that time-related methods can be called.

const SUPPLIES_OFFSET: bool = false

Whether the type can provide offset components, indicating that offset-related methods can be called.

const SUPPLIES_TIMESTAMP: bool = false

Whether the type can provide timestamp components, indicating that timestamp-related methods can be called.

Required Associated Types

type State: Default

The state type used by the provider, allowing for caching of computed values.

Provided Methods

fn day(&self, state: &mut Self::State) -> ru8<1, 31>

Obtain the day of the month.

fn month(&self, state: &mut Self::State) -> Month

Obtain the month of the year.

fn ordinal(&self, state: &mut Self::State) -> ru16<1, 366>

Obtain the ordinal day of the year.

fn weekday(&self, state: &mut Self::State) -> Weekday

Obtain the day of the week.

fn iso_week_number(&self, state: &mut Self::State) -> ru8<1, 53>

Obtain the ISO week number.

fn monday_based_week(&self, state: &mut Self::State) -> ru8<0, 53>

Obtain the Monday-based week number.

fn sunday_based_week(&self, state: &mut Self::State) -> ru8<0, 53>

Obtain the Sunday-based week number.

fn calendar_year(&self, state: &mut Self::State) -> ri32<-999_999, 999_999>

Obtain the calendar year.

fn iso_year(&self, state: &mut Self::State) -> ri32<-999_999, 999_999>

Obtain the ISO week-based year.

fn hour(&self, state: &mut Self::State) -> ru8<0, { _ }>

Obtain the hour within the day.

fn minute(&self, state: &mut Self::State) -> ru8<0, { _ }>

Obtain the minute within the hour.

fn period(&self, state: &mut Self::State) -> Period

Obtain the period of the day (AM/PM).

fn second(&self, state: &mut Self::State) -> ru8<0, { _ }>

Obtain the second within the minute.

fn nanosecond(&self, state: &mut Self::State) -> ru32<0, { _ }>

Obtain the nanosecond within the second.

fn offset_is_negative(&self, state: &mut Self::State) -> bool

Obtain whether the offset is negative.

fn offset_is_utc(&self, state: &mut Self::State) -> bool

Obtain whether the offset is UTC.

fn offset_hour(&self, state: &mut Self::State) -> ri8<-25, 25>

Obtain the hour component of the UTC offset.

fn offset_minute(&self, state: &mut Self::State) -> ri8<{ _ }, { _ }>

Obtain the minute component of the UTC offset.

fn offset_second(&self, state: &mut Self::State) -> ri8<{ _ }, { _ }>

Obtain the second component of the UTC offset.

fn unix_timestamp_seconds(&self, state: &mut Self::State) -> i64

Obtain the Unix timestamp in seconds.

fn unix_timestamp_milliseconds(&self, state: &mut Self::State) -> i64

Obtain the Unix timestamp in milliseconds.

fn unix_timestamp_microseconds(&self, state: &mut Self::State) -> i128

Obtain the Unix timestamp in microseconds.

fn unix_timestamp_nanoseconds(&self, state: &mut Self::State) -> i128

Obtain the Unix timestamp in nanoseconds.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.