time::formatting

Function f64_10_pow_x

fn f64_10_pow_x(x: NonZero<u8>) -> f64

Helper function to obtain 10^x, guaranteeing determinism for x ≤ 9. For these cases, the function optimizes to a lookup table. For x ≥ 10, it falls back to 10_f64.powi(x). The only situation where this would occur is if the user explicitly requests such precision when configuring the ISO 8601 well known format. All other possibilities max out at nine digits.