Trait time::formatting::formattable::sealed::Formattable

pub trait Formattable {
    type Error: From<Error>;
    fn format_into(
        &self, 
        output: &mut impl Write, 
        date: Option<Date>, 
        time: Option<Time>, 
        offset: Option<UtcOffset>
    ) -> Result<usize, Self::Error>;

    fn format(
        &self, 
        date: Option<Date>, 
        time: Option<Time>, 
        offset: Option<UtcOffset>
    ) -> Result<String, Self::Error> { ... }
}

This is supported on crate feature formatting only.

Format the item using a format description, the intended output, and the various components.

Associated Types

type Error: From<Error>

An error that may be returned when formatting.

Loading content...
Required methods

fn format_into(
&self,
output: &mut impl Write,
date: Option<Date>,
time: Option<Time>,
offset: Option<UtcOffset>
) -> Result<usize, Self::Error>

Format the item into the provided output, returning the number of bytes written.

Loading content...
Provided methods

fn format(
&self,
date: Option<Date>,
time: Option<Time>,
offset: Option<UtcOffset>
) -> Result<String, Self::Error>

Format the item directly to a String.

Loading content...