time::formatting::formattable::sealed

Trait Sealed

pub trait Sealed: ComputeMetadata {
    // Required method
    fn format_into<V>(
        &self,
        output: &mut (impl Write + ?Sized),
        value: &V,
        state: &mut V::State,
        _: PrivateMethod,
    ) -> Result<usize, Format>
       where V: ComponentProvider;

    // Provided method
    fn format<V>(
        &self,
        value: &V,
        state: &mut V::State,
        _: PrivateMethod,
    ) -> Result<String, Format>
       where V: ComponentProvider { ... }
}

Format the item using a format description, the intended output, and the various components.

Required Methods

fn format_into<V>( &self, output: &mut (impl Write + ?Sized), value: &V, state: &mut V::State, _: PrivateMethod, ) -> Result<usize, Format>
where V: ComponentProvider,

Format the item into the provided output, returning the number of bytes written.

Provided Methods

fn format<V>( &self, value: &V, state: &mut V::State, _: PrivateMethod, ) -> Result<String, Format>
where V: ComponentProvider,

Format the item directly to a String.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.