time

Module internal_macros

Macros for use within the library. They are not publicly available.

Macros

bug 🔒

unreachable!(), but better.

carry 🔒

Similar to overflowing_add, but returning the number of times that it overflowed. Contained to a certain range and only overflows a maximum number of times.

cascade 🔒

Cascade an out-of-bounds value.

const_try 🔒

Try to unwrap an expression, returning if not possible.

const_try_opt 🔒

Try to unwrap an expression, returning if not possible.

div_floor 🔒

Division of integers, rounding the resulting value towards negative infinity.

ensure_ranged 🔒

Constructs a ranged integer, returning a ComponentRange error if the value is out of range.

try_likely_ok 🔒

Try to unwrap an expression, returning if not possible.