time

Module iter

Iterator types.

Structs

DateIter

An iterator over a range of Dates, yielding each day from start to end inclusive.

MonthIter

An infinite iterator over Months.

Rev

An iterator that yields the elements of an underlying iterator in reverse order.

WeekdayIter

An infinite iterator over Weekdays.