time

Module num_fmt

Formatting utilities for numbers.

These functions are low-level, but are designed to be extremely fast for their designed use cases. They have strict requirements, and may not return the most ergonomic types to avoid unnecessary allocations and copying.

Structs

StackStr 🔒

A string type with a maximum length known at compile time, stored on the stack.

StackTrailingStr 🔒

A string type with a maximum length known at compile time, stored on the stack.

Statics

SINGLE_DIGITS 🔒

SPACE_PADDED_PAIRS 🔒

ZERO_PADDED_PAIRS 🔒

Functions

div_100 🔒

div_rem_1e16 🔒

enc_16lsd 🔒 ⚠

Encodes the 16 least-significant decimals of n into buf[OFFSET..OFFSET + 16].

five_digits_zero_padded 🔒

Obtain three strings which together represent n. The first string is the most significant. Leading zeros are included if the number has fewer than 5 digits. The first string will be empty if n is less than 10,000.

four_digits_space_padded 🔒

Obtain two strings of two ASCII digits each representing n. The first string is the most significant. Leading spaces are included if the number has fewer than 4 digits.

four_digits_zero_padded 🔒

Obtain two strings of two ASCII digits each representing n. The first string is the most significant. Leading zeros are included if the number has fewer than 4 digits.

four_to_six_digits 🔒

Obtain three strings which together represent n. The first string is the most significant. Leading zeros are included if the number has fewer than 4 digits. The first string will be empty if n is less than 10,000.

mulhi 🔒

Multiply unsigned 128 bit integers, return upper 128 bits of the result

one_to_four_digits_no_padding 🔒

Obtain two strings of ASCII digits representing n. The first string is the most significant. No leading zeros or spaces are included.

one_to_three_digits_no_padding 🔒

Obtain two strings of ASCII digits representing n. The first string is most significant. No leading zeros or spaces are included.

one_to_two_digits_no_padding 🔒

Obtain a string of one or two ASCII digits representing n. No leading zeros or spaces are included.

single_digit 🔒

Obtain a string containing a single ASCII digit representing n.

six_digits_zero_padded 🔒

Obtain three strings which together represent n. The first string is the most significant. Leading zeroes are included if the number has fewer than 6 digits.

str_from_raw_parts 🔒 ⚠

Safety

subsecond_from_nanos 🔒

Obtain five strings which together represent n, which is a number of nanoseconds.

three_digits_space_padded 🔒

Obtain two strings of ASCII digits representing n. The first string is the most significant. Leading spaces are included if the number has fewer than 3 digits.

three_digits_zero_padded 🔒

Obtain two strings of ASCII digits representing n. The first string is the most significant. Leading zeros are included if the number has fewer than 3 digits.

truncated_subsecond_from_nanos 🔒

Obtain a string of 1 to 9 ASCII digits representing n, which is a number of nanoseconds.

two_digits_space_padded 🔒

Obtain a string of two ASCII digits representing n. This includes a leading space if n is less than 10.

two_digits_zero_padded 🔒

Obtain a string of two ASCII digits representing n. This includes a leading zero if n is less than 10.

u64_pad_none 🔒

Format a u64 into a string with no padding.

u128_pad_none 🔒

Format a u128 into a string with no padding.

write_one_digit 🔒 ⚠

Write a single digit integer to buf at offset.

write_two_digits 🔒 ⚠

Write a two digit integer to buf at offset and offset + 1.