time

Module offset_date_time

The OffsetDateTime struct and its associated impls.

Structs

OffsetDateTime

A PlainDateTime with a UtcOffset.

Constants

UNIX_EPOCH_JULIAN_DAY 🔒

The Julian day of the Unix epoch.

Functions

raw_to_bits 🔒

Note: This value is explicitly signed, so do not cast this to or treat this as an unsigned integer. Doing so will lead to incorrect results for values with differing signs.