time::parsing

Module combinator

Implementations of the low-level parser combinators.

Modules

rfc 🔒

Combinators for rules as defined in a standard.

Structs

ExactlyNDigits 🔒

Parse an exact number of digits without padding.

Enums

Sign 🔒

The sign of a number.

Functions

any_digit 🔒

Consume exactly one digit.

ascii_char 🔒

Consume exactly one of the provided ASCII characters.

ascii_char_ignore_case 🔒

Consume exactly one of the provided ASCII characters, case-insensitive.

exactly_n_digits_padded 🔒

Consume exactly n digits, returning the numerical value.

n_to_m_digits 🔒

Consume between n and m digits, returning the numerical value.

n_to_m_digits_padded 🔒

Consume between n and m digits, returning the numerical value.

one_or_more 🔒

Consume one of or more instances of the provided parser. The parser must produce the unit value.

one_or_two_digits 🔒

Consume one or two digits, returning the numerical value.

opt 🔒

Optionally consume an input with a given parser.

sign 🔒

Parse a “+” or “-” sign.

zero_or_more 🔒

Consume zero or more instances of the provided parser. The parser must return the unit value.