time::parsing::combinator::rfc

Module iso8601

Rules defined in ISO 8601.

Enums

ExtendedKind 🔒

What kind of format is being parsed. This is used to ensure each part of the format (date, time, offset) is the same kind.

Functions

day 🔒

Parse a day of the month.

dayk 🔒

Parse a day of the week.

dayo 🔒

Parse a day of the year.

decimal_sign 🔒

Parse a “decimal sign”, which is either a comma or a period.

float 🔒

Parse a floating point number as its integer and optional fractional parts.

hour 🔒

Parse the hour.

min 🔒

Parse the minute.

month 🔒

Parse a month.

week 🔒

Parse a week number.

year 🔒

Parse a possibly expanded year.