time::parsing::combinator::rfc

Module rfc2822

Rules defined in RFC 2822.

Constants

DEPTH_LIMIT 🔒

Functions

ccontent 🔒

Consume the ccontent rule.

cfws 🔒

Consume the cfws rule.

comment 🔒

Consume the comment rule.

ctext 🔒

Consume the ctext rule.

fws 🔒

Consume the fws rule.

no_ws_ctl 🔒

Consume the no_ws_ctl rule.

opt_cfws 🔒

Optional cfws rule. This is equivalent to opt(cfws), but is better optimized for the common case where no cfws is present.

opt_cfws_colon_opt_cfws 🔒

Equivalent to opt(cfws), ascii_char::<b':'>, and opt(cfws) called in sequence, but is better optimized for the common case where no cfws is present.

quoted_pair 🔒

Consume the quoted_pair rule.

text 🔒

Consume the text rule.

zone_literal 🔒

Consume an old zone literal, returning the offset in hours.