time::parsing

Module component

Parsing implementations for all Components.

Functions

parse_calendar_year_century_extended_range 🔒

Parse all digits of the calendar-based year except the last two.

parse_calendar_year_century_standard_range 🔒

Parse all digits of the calendar-based year except the last two.

parse_calendar_year_full_extended_range 🔒

Parse the full calendar-based year.

parse_calendar_year_full_standard_range 🔒

Parse the full calendar-based year.

parse_calendar_year_last_two 🔒

Parse the last two digits of the calendar-based year.

parse_day 🔒

Parse the “day” component of a Date.

parse_end 🔒

Parse the end component, which represents the end of input. If any input is remaining and trailing input is prohibited, None is returned. If trailing input is permitted, it is discarded.

parse_hour_12 🔒

Parse the “hour” component of a Time in the 12-hour format.

parse_hour_24 🔒

Parse the “hour” component of a Time in the 24-hour format.

parse_ignore 🔒

Ignore the given number of bytes.

parse_iso_year_century_extended_range 🔒

Parse all digits of the ISO week-based year except the last two.

parse_iso_year_century_standard_range 🔒

Parse all digits of the ISO week-based year except the last two.

parse_iso_year_full_extended_range 🔒

Parse the full ISO-week based year.

parse_iso_year_full_standard_range 🔒

Parse the full ISO-week based year.

parse_iso_year_last_two 🔒

Parse the last two digits of the ISO week-based year.

parse_minute 🔒

Parse the “minute” component of a Time.

parse_month_long 🔒

Parse the “month” component of a Date in the long form (e.g. “January”).

parse_month_numerical 🔒

Parse the “month” component of a Date in the numerical format (e.g. “1” for January).

parse_month_short 🔒

Parse the “month” component of a Date in the abbreviated form (e.g. “Jan”).

parse_offset_hour 🔒

Parse the “hour” component of a UtcOffset.

parse_offset_minute 🔒

Parse the “minute” component of a UtcOffset.

parse_offset_second 🔒

Parse the “second” component of a UtcOffset.

parse_ordinal 🔒

Parse the “ordinal” component of a Date.

parse_period 🔒

Parse the “period” component of a Time. Required if the hour is on a 12-hour clock.

parse_second 🔒

Parse the “second” component of a Time.

parse_subsecond 🔒

Parse the “subsecond” component of a Time.

parse_unix_timestamp_microsecond 🔒

Parse the Unix timestamp component with microsecond precision, returning the value in nanoseconds.

parse_unix_timestamp_millisecond 🔒

Parse the Unix timestamp component with millisecond precision, returning the value in nanoseconds.

parse_unix_timestamp_nanosecond 🔒

Parse the Unix timestamp component with nanosecond precision.

parse_unix_timestamp_second 🔒

Parse the Unix timestamp component with second precision, returning the value in nanoseconds.

parse_week_number_iso 🔒

Parse the “week number” component of a Date, where week 1 starts on the last Monday on or before January 4.

parse_week_number_monday 🔒

Parse the “week number” component of a Date, where week 1 starts on the first Monday of the year.

parse_week_number_sunday 🔒

Parse the “week number” component of a Date, where week 1 starts on the first Sunday of the year.

parse_weekday_long 🔒

Parse the “weekday” component of a Date in the long form (e.g. “Monday”).

parse_weekday_monday 🔒

Parse the weekday component of a Date in the numerical format, where Monday is the first day of the week.`

parse_weekday_short 🔒

Parse the “weekday” component of a Date in the abbreviated form (e.g. “Mon”).

parse_weekday_sunday 🔒

Parse the weekday component of a Date in the numerical format, where Sunday is the first day of the week.`