time::parsing::parsable::sealed

Trait Sealed

pub trait Sealed {
    // Required method
    fn parse_into<'a>(
        &self,
        input: &'a [u8],
        parsed: &mut Parsed,
        _: PrivateMethod,
    ) -> Result<&'a [u8], Parse>;

    // Provided methods
    fn parse_internal(
        &self,
        input: &[u8],
        defaults: Option<Parsed>,
        _: PrivateMethod,
    ) -> Result<Parsed, Parse> { ... }
    fn parse_date(
        &self,
        input: &[u8],
        defaults: Option<Parsed>,
        _: PrivateMethod,
    ) -> Result<Date, Parse> { ... }
    fn parse_time(
        &self,
        input: &[u8],
        defaults: Option<Parsed>,
        _: PrivateMethod,
    ) -> Result<Time, Parse> { ... }
    fn parse_offset(
        &self,
        input: &[u8],
        defaults: Option<Parsed>,
        _: PrivateMethod,
    ) -> Result<UtcOffset, Parse> { ... }
    fn parse_plain_date_time(
        &self,
        input: &[u8],
        defaults: Option<Parsed>,
        _: PrivateMethod,
    ) -> Result<PlainDateTime, Parse> { ... }
    fn parse_utc_date_time(
        &self,
        input: &[u8],
        defaults: Option<Parsed>,
        _: PrivateMethod,
    ) -> Result<UtcDateTime, Parse> { ... }
    fn parse_offset_date_time(
        &self,
        input: &[u8],
        defaults: Option<Parsed>,
        _: PrivateMethod,
    ) -> Result<OffsetDateTime, Parse> { ... }
    fn parse_timestamp(
        &self,
        input: &[u8],
        defaults: Option<Parsed>,
        _: PrivateMethod,
    ) -> Result<Timestamp, Parse> { ... }
}

Parse the item using a format description and an input.

Required Methods

fn parse_into<'a>( &self, input: &'a [u8], parsed: &mut Parsed, _: PrivateMethod, ) -> Result<&'a [u8], Parse>

Parse the item into the provided Parsed struct.

This method can be used to parse a single component without parsing the full value.

Provided Methods

fn parse_internal( &self, input: &[u8], defaults: Option<Parsed>, _: PrivateMethod, ) -> Result<Parsed, Parse>

Parse the items into a Parsed struct, using the provided defaults for any components that are not present in the input.

This method can only be used to parse a complete value of a type. If any characters remain after parsing, an error will be returned.

fn parse_date( &self, input: &[u8], defaults: Option<Parsed>, _: PrivateMethod, ) -> Result<Date, Parse>

Parse a Date from the format description.

fn parse_time( &self, input: &[u8], defaults: Option<Parsed>, _: PrivateMethod, ) -> Result<Time, Parse>

Parse a Time from the format description.

fn parse_offset( &self, input: &[u8], defaults: Option<Parsed>, _: PrivateMethod, ) -> Result<UtcOffset, Parse>

Parse a UtcOffset from the format description.

fn parse_plain_date_time( &self, input: &[u8], defaults: Option<Parsed>, _: PrivateMethod, ) -> Result<PlainDateTime, Parse>

Parse a PlainDateTime from the format description.

fn parse_utc_date_time( &self, input: &[u8], defaults: Option<Parsed>, _: PrivateMethod, ) -> Result<UtcDateTime, Parse>

Parse a UtcDateTime from the format description.

fn parse_offset_date_time( &self, input: &[u8], defaults: Option<Parsed>, _: PrivateMethod, ) -> Result<OffsetDateTime, Parse>

Parse a OffsetDateTime from the format description.

fn parse_timestamp( &self, input: &[u8], defaults: Option<Parsed>, _: PrivateMethod, ) -> Result<Timestamp, Parse>

Parse a Timestamp from the format description.