Module time::serde

This is supported on crate feature serde only.

Differential formats for serde.

Modules

timestamp
Treat an OffsetDateTime as a Unix timestamp for the purposes of serde.

Constants

DATE_FORMAT
The format used when serializing and deserializing a human-readable Date.

OFFSET_DATE_TIME_FORMAT
The format used when serializing and deserializing a human-readable OffsetDateTime.

PRIMITIVE_DATE_TIME_FORMAT
The format used when serializing and deserializing a human-readable PrimitiveDateTime.

TIME_FORMAT
The format used when serializing and deserializing a human-readable Time.

UTC_OFFSET_FORMAT
The format used when serializing and deserializing a human-readable UtcOffset.