time::signed_duration

Macro try_from_secs

macro_rules! try_from_secs {
    (
        secs = $secs: expr,
        mantissa_bits = $mant_bits: literal,
        exponent_bits = $exp_bits: literal,
        offset = $offset: literal,
        bits_ty = $bits_ty:ty,
        bits_ty_signed = $bits_ty_signed:ty,
        double_ty = $double_ty:ty,
        float_ty = $float_ty:ty,
    ) => { ... };
}

This is adapted from the std implementation, which uses mostly bit operations to ensure the highest precision:

Changes from std are marked and explained below.