Struct time::Instant

#[repr(transparent)]
pub struct Instant(pub StdInstant);

A measurement of a monotonically non-decreasing clock. Opaque and useful only with Duration.

Instants are always guaranteed to be no less than any previously measured instant when created, and are often useful for tasks such as measuring benchmarks or timing how long an operation takes.

Note, however, that instants are not guaranteed to be steady. In other words, each tick of the underlying clock may not be the same length (e.g. some seconds may be longer than others). An instant may jump forwards or experience time dilation (slow down or speed up), but it will never go backwards.

Instants are opaque types that can only be compared to one another. There is no method to get “the number of seconds” from an instant. Instead, it only allows measuring the duration between two instants (or comparing two instants).

This implementation allows for operations with signed Durations, but is otherwise identical to [std::time::Instant].

Tuple Fields
0: StdInstant
Implementations

impl Instant

pub fn now() -> Self

Returns an Instant corresponding to “now”.

println!("{:?}", Instant::now());
Run

pub fn elapsed(self) -> Duration

Returns the amount of time elapsed since this instant was created. The duration will always be nonnegative if the instant is not synthetically created.

let instant = Instant::now();
thread::sleep(1.std_milliseconds());
assert!(instant.elapsed() >= 1.milliseconds());
Run

pub fn checked_add(self, duration: Duration) -> Option<Self>

Returns Some(t) where t is the time self + duration if t can be represented as Instant (which means it’s inside the bounds of the underlying data structure), None otherwise.

let now = Instant::now();
assert_eq!(now.checked_add(5.seconds()), Some(now + 5.seconds()));
assert_eq!(now.checked_add((-5).seconds()), Some(now + (-5).seconds()));
Run

pub fn checked_sub(self, duration: Duration) -> Option<Self>

Returns Some(t) where t is the time self - duration if t can be represented as Instant (which means it’s inside the bounds of the underlying data structure), None otherwise.

let now = Instant::now();
assert_eq!(now.checked_sub(5.seconds()), Some(now - 5.seconds()));
assert_eq!(now.checked_sub((-5).seconds()), Some(now - (-5).seconds()));
Run

pub const fn into_inner(self) -> StdInstant

Obtain the inner [std::time::Instant].

let now = Instant::now();
assert_eq!(now.into_inner(), now.0);
Run