Struct time::OffsetDateTime

pub struct OffsetDateTime { /* fields omitted */ }

A DateTime with a UtcOffset.

For equality, comparisons, and hashing, calculations are performed using the Unix timestamp.

Methods

impl OffsetDateTime

pub fn now() -> Self

Create a new OffsetDateTime with the current date and time (UTC).

assert!(OffsetDateTime::now().year() >= 2019);
assert_eq!(OffsetDateTime::now().offset(), UtcOffset::UTC);

pub const fn to_offset(self, offset: UtcOffset) -> Self

Convert the OffsetDateTime from the current UtcOffset to the provided UtcOffset.

assert_eq!(
    Date::from_ymd(2000, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .to_offset(UtcOffset::hours(-1))
        .year(),
    1999,
);

pub const fn unix_epoch() -> Self

Midnight, 1 January, 1970 (UTC).

assert_eq!(
    OffsetDateTime::unix_epoch(),
    Date::from_ymd(1970, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC),
);

pub fn from_unix_timestamp(timestamp: i64) -> Self

Create an OffsetDateTime from the provided Unix timestamp.

assert_eq!(
    OffsetDateTime::from_unix_timestamp(0),
    OffsetDateTime::unix_epoch(),
);
assert_eq!(
    OffsetDateTime::from_unix_timestamp(1_546_300_800),
    Date::from_ymd(2019, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC),
);

pub const fn offset(self) -> UtcOffset

Get the UtcOffset of the OffsetDateTime.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms(0, 0, 0)
        .using_offset(UtcOffset::UTC)
        .offset(),
    UtcOffset::UTC,
);
assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms(0, 0, 0)
        .using_offset(UtcOffset::hours(1))
        .offset(),
    UtcOffset::hours(1),
);

pub fn timestamp(self) -> i64

Get the Unix timestamp representing the OffsetDateTime.

assert_eq!(
    DateTime::unix_epoch()
        .using_offset(UtcOffset::UTC)
        .timestamp(),
    0,
);
assert_eq!(
    DateTime::unix_epoch()
        .using_offset(UtcOffset::hours(-1))
        .timestamp(),
    3_600,
);

pub fn date(self) -> Date

Get the Date in the stored offset of the OffsetDateTime.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .date(),
    Date::from_ymd(2019, 1, 1),
);
assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .midnight()
        .using_offset(UtcOffset::hours(-1))
        .date(),
    Date::from_ymd(2018, 12, 31),
);

pub fn time(self) -> Time

Get the Time in the stored offset of the OffsetDateTime.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .time(),
    Time::from_hms(0, 0, 0),
);
assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .midnight()
        .using_offset(UtcOffset::hours(-1))
        .time(),
    Time::from_hms(23, 0, 0),
);

pub fn year(self) -> i32

Get the year of the date in the stored offset.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .year(),
    2019,
);
assert_eq!(
    Date::from_ymd(2019, 12, 31)
        .with_hms(23, 0, 0)
        .using_offset(UtcOffset::UTC)
        .to_offset(UtcOffset::hours(1))
        .year(),
    2020,
);
assert_eq!(
    Date::from_ymd(2020, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .year(),
    2020,
);

pub fn month(self) -> u8

Get the month of the date in the stored offset. If fetching both the month and day, use [OffsetDateTime::month_day] instead.

The returned value will always be in the range 1..=12.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .month(),
    1,
);
assert_eq!(
    Date::from_ymd(2019, 12, 31)
        .with_hms(23, 0, 0)
        .using_offset(UtcOffset::hours(1))
        .month(),
    1,
);

pub fn day(self) -> u8

Get the day of the date in the stored offset. If fetching both the month and day, use [OffsetDateTime::month_day] instead.

The returned value will always be in the range 1..=31.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .day(),
    1,
);
assert_eq!(
    Date::from_ymd(2019, 12, 31)
        .with_hms(23, 0, 0)
        .using_offset(UtcOffset::hours(1))
        .day(),
    1,
);

pub fn month_day(self) -> (u8, u8)

Get the month and day of the date in the stored offset.

The month component will always be in the range 1..=12; the day component in 1..=31.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .month_day(),
    (1, 1),
);
assert_eq!(
    Date::from_ymd(2019, 12, 31)
        .with_hms(23, 0, 0)
        .using_offset(UtcOffset::hours(1))
        .month_day(),
    (1, 1),
);

pub fn ordinal(self) -> u16

Get the day of the year of the date in the stored offset.

The returned value will always be in the range 1..=366.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .ordinal(),
    1,
);
assert_eq!(
    Date::from_ymd(2019, 12, 31)
        .with_hms(23, 0, 0)
        .using_offset(UtcOffset::hours(1))
        .ordinal(),
    1,
);

pub fn week(self) -> u8

Get the ISO week number of the date in the stored offset.

The returned value will always be in the range 1..=53.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .week(),
    1,
);
assert_eq!(
    Date::from_ymd(2020, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .week(),
    1,
);
assert_eq!(
    Date::from_ymd(2020, 12, 31)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .week(),
    53,
);
assert_eq!(
    Date::from_ymd(2021, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .week(),
    53,
);

pub fn weekday(self) -> Weekday

Get the weekday of the date in the stored offset.

This current uses Zeller's congruence internally.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .weekday(),
    Tuesday,
);
assert_eq!(
    Date::from_ymd(2019, 2, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .weekday(),
    Friday,
);
assert_eq!(
    Date::from_ymd(2019, 3, 1)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .weekday(),
    Friday,
);

pub fn hour(self) -> u8

Returns the clock hour in the stored offset.

The returned value will always be in the range 0..=23.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms(0, 0, 0)
        .using_offset(UtcOffset::UTC)
        .hour(),
    0,
);
assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms(23, 59, 59)
        .using_offset(UtcOffset::hours(-2))
        .hour(),
    21,
);

pub fn minute(self) -> u8

Returns the minute within the hour in the stored offset.

The returned value will always be in the range 0..60.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms(0, 0, 0)
        .using_offset(UtcOffset::UTC)
        .minute(),
    0,
);
assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms(23, 59, 59)
        .using_offset(UtcOffset::minutes(30))
        .minute(),
    29,
);

pub fn second(self) -> u8

Returns the second within the minute in the stored offset.

The returned value will always be in the range 0..60.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms(0, 0, 0)
        .using_offset(UtcOffset::UTC)
        .second(),
    0,
);
assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms(23, 59, 59)
        .using_offset(UtcOffset::seconds(30))
        .second(),
    29,
);

pub fn millisecond(self) -> u16

Return the milliseconds within the second in the stored offset.

The returned value will always be in the range 0..1_000.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms_milli(0, 0, 0, 0)
        .using_offset(UtcOffset::UTC)
        .millisecond(),
    0,
);
assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms_milli(23, 59, 59, 999)
        .using_offset(UtcOffset::UTC)
        .millisecond(),
    999,
);

pub fn microsecond(self) -> u32

Return the microseconds within the second in the stored offset.

The returned value will always be in the range 0..1_000_000.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms_micro(0, 0, 0, 0)
        .using_offset(UtcOffset::UTC)
        .microsecond(),
    0,
);
assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms_micro(23, 59, 59, 999_999)
        .using_offset(UtcOffset::UTC)
        .microsecond(),
    999_999,
);

pub fn nanosecond(self) -> u32

Return the nanoseconds within the second in the stored offset.

The returned value will always be in the range 0..1_000_000_000.

assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms_nano(0, 0, 0, 0)
        .using_offset(UtcOffset::UTC)
        .nanosecond(),
    0,
);
assert_eq!(
    Date::from_ymd(2019, 1, 1)
        .with_hms_nano(23, 59, 59, 999_999_999)
        .using_offset(UtcOffset::UTC)
        .nanosecond(),
    999_999_999,
);

impl OffsetDateTime

Methods that allow formatting the OffsetDateTime.

pub fn format(self, format: &str) -> String

Format the OffsetDateTime using the provided string. As no language is specified, English is used.

assert_eq!(
    Date::from_ymd(2019, 1, 2)
        .midnight()
        .using_offset(UtcOffset::UTC)
        .format("%F %r %z"),
    "2019-01-02 12:00:00 am +0000",
);

pub fn format_language(self, format: &str, language: Language) -> String

Format the OffsetDateTime using the provided string and language.

assert_eq!(
    Date::from_ymd(2019, 1, 2)
        .midnight()
        .using_offset(UtcOffset::hours(-2))
        .format_language("%c %z", Language::en),
    "Tue Jan 1 22:00:00 2019 -0200",
);
assert_eq!(
    Date::from_ymd(2019, 1, 2)
        .midnight()
        .using_offset(UtcOffset::hours(2))
        .format_language("%c %z", Language::es),
    "Mi enero 2 02:00:00 2019 +0200",
);