Struct time::PreciseTime

pub struct PreciseTime(_);

An opaque structure representing a moment in time.

The only operation that can be performed on a PreciseTime is the calculation of the Duration of time that lies between them.

Examples

Repeatedly call a function for 1 second:

use std::time::Duration;
use time::PreciseTime;

let start = PreciseTime::now();

while start.to(PreciseTime::now()) < Duration::seconds(1) {
    do_some_work();
}

Methods

impl PreciseTime

fn now() -> PreciseTime

Returns a PreciseTime representing the current moment in time.

fn to(&self, later: PreciseTime) -> Duration

Returns a Duration representing the span of time from the value of self to the value of later.

Notes

If later represents a time before self, the result of this method is unspecified.

If later represents a time more than 293 years after self, the result of this method is unspecified.