time

Struct SignedDuration

pub struct SignedDuration { /* private fields */ }

A span of time with nanosecond precision.

Each SignedDuration is composed of a whole number of seconds and a fractional part represented in nanoseconds.

This implementation allows for negative durations, unlike core::time::Duration.

Implementations

impl SignedDuration

pub const ZERO: Self

Equivalent to 0.seconds().

assert_eq!(SignedDuration::ZERO, 0.seconds());

pub const NANOSECOND: Self

Equivalent to 1.nanoseconds().

assert_eq!(SignedDuration::NANOSECOND, 1.nanoseconds());

pub const MICROSECOND: Self

Equivalent to 1.microseconds().

assert_eq!(SignedDuration::MICROSECOND, 1.microseconds());

pub const MILLISECOND: Self

Equivalent to 1.milliseconds().

assert_eq!(SignedDuration::MILLISECOND, 1.milliseconds());

pub const SECOND: Self

Equivalent to 1.seconds().

assert_eq!(SignedDuration::SECOND, 1.seconds());

pub const MINUTE: Self

Equivalent to 1.minutes().

assert_eq!(SignedDuration::MINUTE, 1.minutes());

pub const HOUR: Self

Equivalent to 1.hours().

assert_eq!(SignedDuration::HOUR, 1.hours());

pub const DAY: Self

Equivalent to 1.days().

assert_eq!(SignedDuration::DAY, 1.days());

pub const WEEK: Self

Equivalent to 1.weeks().

assert_eq!(SignedDuration::WEEK, 1.weeks());

pub const MIN: Self

The minimum possible duration. Adding any negative duration to this will cause an overflow.

pub const MAX: Self

The maximum possible duration. Adding any positive duration to this will cause an overflow.

pub const fn is_zero(self) -> bool

Check if a duration is exactly zero.

assert!(0.seconds().is_zero());
assert!(!1.nanoseconds().is_zero());

pub const fn is_negative(self) -> bool

Check if a duration is negative.

assert!((-1).seconds().is_negative());
assert!(!0.seconds().is_negative());
assert!(!1.seconds().is_negative());

pub const fn is_positive(self) -> bool

Check if a duration is positive.

assert!(1.seconds().is_positive());
assert!(!0.seconds().is_positive());
assert!(!(-1).seconds().is_positive());

pub const fn abs(self) -> Self

Get the absolute value of the duration.

This method saturates the returned value if it would otherwise overflow.

assert_eq!(1.seconds().abs(), 1.seconds());
assert_eq!(0.seconds().abs(), 0.seconds());
assert_eq!((-1).seconds().abs(), 1.seconds());

pub const fn unsigned_abs(self) -> StdDuration

Convert the existing SignedDuration to a std::time::Duration and its sign. This returns a std::time::Duration and does not saturate the returned value (unlike SignedDuration::abs).

assert_eq!(1.seconds().unsigned_abs(), 1.std_seconds());
assert_eq!(0.seconds().unsigned_abs(), 0.std_seconds());
assert_eq!((-1).seconds().unsigned_abs(), 1.std_seconds());

pub const fn new(seconds: i64, nanoseconds: i32) -> Self

Create a new SignedDuration with the provided seconds and nanoseconds. If nanoseconds is at least ±109, it will wrap to the number of seconds.

assert_eq!(SignedDuration::new(1, 0), 1.seconds());
assert_eq!(SignedDuration::new(-1, 0), (-1).seconds());
assert_eq!(SignedDuration::new(1, 2_000_000_000), 3.seconds());

Panics

This may panic if an overflow occurs.

pub const fn weeks(weeks: i64) -> Self

Create a new SignedDuration with the given number of weeks. Equivalent to SignedDuration::seconds(weeks * 604_800).

assert_eq!(SignedDuration::weeks(1), 604_800.seconds());

Panics

This may panic if an overflow occurs.

pub const fn days(days: i64) -> Self

Create a new SignedDuration with the given number of days. Equivalent to SignedDuration::seconds(days * 86_400).

assert_eq!(SignedDuration::days(1), 86_400.seconds());

Panics

This may panic if an overflow occurs.

pub const fn hours(hours: i64) -> Self

Create a new SignedDuration with the given number of hours. Equivalent to SignedDuration::seconds(hours * 3_600).

assert_eq!(SignedDuration::hours(1), 3_600.seconds());

Panics

This may panic if an overflow occurs.

pub const fn minutes(minutes: i64) -> Self

Create a new SignedDuration with the given number of minutes. Equivalent to SignedDuration::seconds(minutes * 60).

assert_eq!(SignedDuration::minutes(1), 60.seconds());

Panics

This may panic if an overflow occurs.

pub const fn seconds(seconds: i64) -> Self

Create a new SignedDuration with the given number of seconds.

assert_eq!(SignedDuration::seconds(1), 1_000.milliseconds());

pub const fn seconds_f64(seconds: f64) -> Self

Creates a new SignedDuration from the specified number of seconds represented as f64.

assert_eq!(SignedDuration::seconds_f64(0.5), 0.5.seconds());
assert_eq!(SignedDuration::seconds_f64(-0.5), (-0.5).seconds());

Panics

This may panic if seconds is NaN or overflows the representable range of SignedDuration.

pub const fn seconds_f32(seconds: f32) -> Self

Creates a new SignedDuration from the specified number of seconds represented as f32.

assert_eq!(SignedDuration::seconds_f32(0.5), 0.5.seconds());
assert_eq!(SignedDuration::seconds_f32(-0.5), (-0.5).seconds());

Panics

This may panic if seconds is NaN or overflows the representable range of SignedDuration.

pub const fn saturating_seconds_f64(seconds: f64) -> Self

Creates a new SignedDuration from the specified number of seconds represented as f64. Any values that are out of bounds are saturated at the minimum or maximum respectively. NaN gets turned into a SignedDuration of 0 seconds.

assert_eq!(SignedDuration::saturating_seconds_f64(0.5), 0.5.seconds());
assert_eq!(
    SignedDuration::saturating_seconds_f64(-0.5),
    (-0.5).seconds()
);
assert_eq!(
    SignedDuration::saturating_seconds_f64(f64::NAN),
    SignedDuration::new(0, 0),
);
assert_eq!(
    SignedDuration::saturating_seconds_f64(f64::NEG_INFINITY),
    SignedDuration::MIN,
);
assert_eq!(
    SignedDuration::saturating_seconds_f64(f64::INFINITY),
    SignedDuration::MAX,
);

pub const fn saturating_seconds_f32(seconds: f32) -> Self

Creates a new SignedDuration from the specified number of seconds represented as f32. Any values that are out of bounds are saturated at the minimum or maximum respectively. NaN gets turned into a SignedDuration of 0 seconds.

assert_eq!(SignedDuration::saturating_seconds_f32(0.5), 0.5.seconds());
assert_eq!(
    SignedDuration::saturating_seconds_f32(-0.5),
    (-0.5).seconds()
);
assert_eq!(
    SignedDuration::saturating_seconds_f32(f32::NAN),
    SignedDuration::new(0, 0),
);
assert_eq!(
    SignedDuration::saturating_seconds_f32(f32::NEG_INFINITY),
    SignedDuration::MIN,
);
assert_eq!(
    SignedDuration::saturating_seconds_f32(f32::INFINITY),
    SignedDuration::MAX,
);

pub const fn checked_seconds_f64(seconds: f64) -> Option<Self>

Creates a new SignedDuration from the specified number of seconds represented as f64. Returns None if the SignedDuration can’t be represented.

assert_eq!(
    SignedDuration::checked_seconds_f64(0.5),
    Some(0.5.seconds())
);
assert_eq!(
    SignedDuration::checked_seconds_f64(-0.5),
    Some((-0.5).seconds())
);
assert_eq!(SignedDuration::checked_seconds_f64(f64::NAN), None);
assert_eq!(SignedDuration::checked_seconds_f64(f64::NEG_INFINITY), None);
assert_eq!(SignedDuration::checked_seconds_f64(f64::INFINITY), None);

pub const fn checked_seconds_f32(seconds: f32) -> Option<Self>

Creates a new SignedDuration from the specified number of seconds represented as f32. Returns None if the SignedDuration can’t be represented.

assert_eq!(
    SignedDuration::checked_seconds_f32(0.5),
    Some(0.5.seconds())
);
assert_eq!(
    SignedDuration::checked_seconds_f32(-0.5),
    Some((-0.5).seconds())
);
assert_eq!(SignedDuration::checked_seconds_f32(f32::NAN), None);
assert_eq!(SignedDuration::checked_seconds_f32(f32::NEG_INFINITY), None);
assert_eq!(SignedDuration::checked_seconds_f32(f32::INFINITY), None);

pub const fn milliseconds(milliseconds: i64) -> Self

Create a new SignedDuration with the given number of milliseconds.

assert_eq!(SignedDuration::milliseconds(1), 1_000.microseconds());
assert_eq!(SignedDuration::milliseconds(-1), (-1_000).microseconds());

pub const fn microseconds(microseconds: i64) -> Self

Create a new SignedDuration with the given number of microseconds.

assert_eq!(SignedDuration::microseconds(1), 1_000.nanoseconds());
assert_eq!(SignedDuration::microseconds(-1), (-1_000).nanoseconds());

pub const fn nanoseconds(nanoseconds: i64) -> Self

Create a new SignedDuration with the given number of nanoseconds.

assert_eq!(SignedDuration::nanoseconds(1), 1.microseconds() / 1_000);
assert_eq!(SignedDuration::nanoseconds(-1), (-1).microseconds() / 1_000);

pub const fn nanoseconds_i128(nanoseconds: i128) -> Self

Create a new SignedDuration with the given number of nanoseconds.

assert_eq!(
    SignedDuration::nanoseconds_i128(1_234_567_890),
    1.seconds() + 234_567_890.nanoseconds()
);

Panics

This may panic if an overflow occurs. This may happen because the input range cannot be fully mapped to the output.

pub const fn whole_weeks(self) -> i64

Get the number of whole weeks in the duration.

assert_eq!(1.weeks().whole_weeks(), 1);
assert_eq!((-1).weeks().whole_weeks(), -1);
assert_eq!(6.days().whole_weeks(), 0);
assert_eq!((-6).days().whole_weeks(), 0);

pub const fn whole_days(self) -> i64

Get the number of whole days in the duration.

assert_eq!(1.days().whole_days(), 1);
assert_eq!((-1).days().whole_days(), -1);
assert_eq!(23.hours().whole_days(), 0);
assert_eq!((-23).hours().whole_days(), 0);

pub const fn whole_hours(self) -> i64

Get the number of whole hours in the duration.

assert_eq!(1.hours().whole_hours(), 1);
assert_eq!((-1).hours().whole_hours(), -1);
assert_eq!(59.minutes().whole_hours(), 0);
assert_eq!((-59).minutes().whole_hours(), 0);

pub const fn whole_minutes(self) -> i64

Get the number of whole minutes in the duration.

assert_eq!(1.minutes().whole_minutes(), 1);
assert_eq!((-1).minutes().whole_minutes(), -1);
assert_eq!(59.seconds().whole_minutes(), 0);
assert_eq!((-59).seconds().whole_minutes(), 0);

pub const fn whole_seconds(self) -> i64

Get the number of whole seconds in the duration.

assert_eq!(1.seconds().whole_seconds(), 1);
assert_eq!((-1).seconds().whole_seconds(), -1);
assert_eq!(1.minutes().whole_seconds(), 60);
assert_eq!((-1).minutes().whole_seconds(), -60);

pub const fn as_seconds_f64(self) -> f64

Get the number of fractional seconds in the duration.

assert_eq!(1.5.seconds().as_seconds_f64(), 1.5);
assert_eq!((-1.5).seconds().as_seconds_f64(), -1.5);

pub const fn as_seconds_f32(self) -> f32

Get the number of fractional seconds in the duration.

assert_eq!(1.5.seconds().as_seconds_f32(), 1.5);
assert_eq!((-1.5).seconds().as_seconds_f32(), -1.5);

pub const fn whole_milliseconds(self) -> i128

Get the number of whole milliseconds in the duration.

assert_eq!(1.seconds().whole_milliseconds(), 1_000);
assert_eq!((-1).seconds().whole_milliseconds(), -1_000);
assert_eq!(1.milliseconds().whole_milliseconds(), 1);
assert_eq!((-1).milliseconds().whole_milliseconds(), -1);

pub const fn subsec_milliseconds(self) -> i16

Get the number of milliseconds past the number of whole seconds.

Always in the range -999..=999.

assert_eq!(1.4.seconds().subsec_milliseconds(), 400);
assert_eq!((-1.4).seconds().subsec_milliseconds(), -400);

pub const fn whole_microseconds(self) -> i128

Get the number of whole microseconds in the duration.

assert_eq!(1.milliseconds().whole_microseconds(), 1_000);
assert_eq!((-1).milliseconds().whole_microseconds(), -1_000);
assert_eq!(1.microseconds().whole_microseconds(), 1);
assert_eq!((-1).microseconds().whole_microseconds(), -1);

pub const fn subsec_microseconds(self) -> i32

Get the number of microseconds past the number of whole seconds.

Always in the range -999_999..=999_999.

assert_eq!(1.0004.seconds().subsec_microseconds(), 400);
assert_eq!((-1.0004).seconds().subsec_microseconds(), -400);

pub const fn whole_nanoseconds(self) -> i128

Get the number of nanoseconds in the duration.

assert_eq!(1.microseconds().whole_nanoseconds(), 1_000);
assert_eq!((-1).microseconds().whole_nanoseconds(), -1_000);
assert_eq!(1.nanoseconds().whole_nanoseconds(), 1);
assert_eq!((-1).nanoseconds().whole_nanoseconds(), -1);

pub const fn subsec_nanoseconds(self) -> i32

Get the number of nanoseconds past the number of whole seconds.

The returned value will always be in the range -999_999_999..=999_999_999.

assert_eq!(1.000_000_400.seconds().subsec_nanoseconds(), 400);
assert_eq!((-1.000_000_400).seconds().subsec_nanoseconds(), -400);

pub const fn checked_add(self, rhs: Self) -> Option<Self>

Computes self + rhs, returning None if an overflow occurred.

assert_eq!(5.seconds().checked_add(5.seconds()), Some(10.seconds()));
assert_eq!(SignedDuration::MAX.checked_add(1.nanoseconds()), None);
assert_eq!((-5).seconds().checked_add(5.seconds()), Some(0.seconds()));

pub const fn checked_sub(self, rhs: Self) -> Option<Self>

Computes self - rhs, returning None if an overflow occurred.

assert_eq!(
    5.seconds().checked_sub(5.seconds()),
    Some(SignedDuration::ZERO)
);
assert_eq!(SignedDuration::MIN.checked_sub(1.nanoseconds()), None);
assert_eq!(5.seconds().checked_sub(10.seconds()), Some((-5).seconds()));

pub const fn checked_mul(self, rhs: i32) -> Option<Self>

Computes self * rhs, returning None if an overflow occurred.

assert_eq!(5.seconds().checked_mul(2), Some(10.seconds()));
assert_eq!(5.seconds().checked_mul(-2), Some((-10).seconds()));
assert_eq!(5.seconds().checked_mul(0), Some(0.seconds()));
assert_eq!(SignedDuration::MAX.checked_mul(2), None);
assert_eq!(SignedDuration::MIN.checked_mul(2), None);

pub const fn checked_div(self, rhs: i32) -> Option<Self>

Computes self / rhs, returning None if rhs == 0 or if the result would overflow.

assert_eq!(10.seconds().checked_div(2), Some(5.seconds()));
assert_eq!(10.seconds().checked_div(-2), Some((-5).seconds()));
assert_eq!(1.seconds().checked_div(0), None);

pub const fn checked_neg(self) -> Option<Self>

Computes -self, returning None if the result would overflow.

assert_eq!(5.seconds().checked_neg(), Some((-5).seconds()));
assert_eq!(SignedDuration::MIN.checked_neg(), None);

pub const fn saturating_add(self, rhs: Self) -> Self

Computes self + rhs, saturating if an overflow occurred.

assert_eq!(5.seconds().saturating_add(5.seconds()), 10.seconds());
assert_eq!(
    SignedDuration::MAX.saturating_add(1.nanoseconds()),
    SignedDuration::MAX
);
assert_eq!(
    SignedDuration::MIN.saturating_add((-1).nanoseconds()),
    SignedDuration::MIN
);
assert_eq!(
    (-5).seconds().saturating_add(5.seconds()),
    SignedDuration::ZERO
);

pub const fn saturating_sub(self, rhs: Self) -> Self

Computes self - rhs, saturating if an overflow occurred.

assert_eq!(
    5.seconds().saturating_sub(5.seconds()),
    SignedDuration::ZERO
);
assert_eq!(
    SignedDuration::MIN.saturating_sub(1.nanoseconds()),
    SignedDuration::MIN
);
assert_eq!(
    SignedDuration::MAX.saturating_sub((-1).nanoseconds()),
    SignedDuration::MAX
);
assert_eq!(5.seconds().saturating_sub(10.seconds()), (-5).seconds());

pub const fn saturating_mul(self, rhs: i32) -> Self

Computes self * rhs, saturating if an overflow occurred.

assert_eq!(5.seconds().saturating_mul(2), 10.seconds());
assert_eq!(5.seconds().saturating_mul(-2), (-10).seconds());
assert_eq!(5.seconds().saturating_mul(0), SignedDuration::ZERO);
assert_eq!(SignedDuration::MAX.saturating_mul(2), SignedDuration::MAX);
assert_eq!(SignedDuration::MIN.saturating_mul(2), SignedDuration::MIN);
assert_eq!(SignedDuration::MAX.saturating_mul(-2), SignedDuration::MIN);
assert_eq!(SignedDuration::MIN.saturating_mul(-2), SignedDuration::MAX);