Struct time::SteadyTime

pub struct SteadyTime(_);

A structure representing a moment in time.

SteadyTimes are generated by a "steady" clock, that is, a clock which never experiences discontinuous jumps and for which time always flows at the same rate.

Examples

Repeatedly call a function for 1 second:

let start = SteadyTime::now();

while SteadyTime::now() - start < Duration::seconds(1) {
    do_some_work();
}

Methods

impl SteadyTime

fn now() -> SteadyTime

Returns a SteadyTime representing the current moment in time.