Struct time::Timespec

pub struct Timespec {
    pub sec: i64,
    pub nsec: i32,
}

A record specifying a time value in seconds and nanoseconds, where nanoseconds represent the offset from the given second.

For example a timespec of 1.2 seconds after the beginning of the epoch would be represented as {sec: 1, nsec: 200000000}.

Fields
sec: i64 nsec: i32
Methods

impl Timespec

pub fn new(sec: i64, nsec: i32) -> Timespec