time

Module time

The Time struct and its associated impls.

Structs

Time

The clock time within a given date. Nanosecond precision.

Enums

Padding 🔒

By explicitly inserting this enum where padding is expected, the compiler is able to better perform niche value optimization.

Type Aliases

Hours 🔒

The type of the hour field of Time.

Minutes 🔒

The type of the minute field of Time.

Nanoseconds 🔒

The type of the nanosecond field of Time.

Seconds 🔒

The type of the second field of Time.