Trait time::ClockSource

pub trait ClockSource {
    type ClockType: ClockType;

    fn now() -> <Self::ClockType as ClockType>::Unit;
    fn unit_to_duration(unit: <Self::ClockType as ClockType>::Unit) -> Duration;
    fn duration_to_unit(
        duration: Duration
    ) -> <Self::ClockType as ClockType>::Unit;
}

A clock source.

Required Associated Types

type ClockType: ClockType

The type of clock (either Monotonic or WallTime).

Required Methods

fn now() -> <Self::ClockType as ClockType>::Unit

The current time according to the clock.

Monotonic clocks return an Instant whereas wall time clocks return a Duration signifying the time since 12:00am January 1st 1970 (i.e. Unix time).

fn unit_to_duration(unit: <Self::ClockType as ClockType>::Unit) -> Duration

Converts a ClockType::Unit into a Duration.

Monotonic clocks should just return the ClockType::Unit.

fn duration_to_unit(duration: Duration) -> <Self::ClockType as ClockType>::Unit

Converts a Duration into a ClockType::Unit.

Monotonic clocks should just return the Duration.