time

Type Alias Duration

pub type Duration = SignedDuration;

A SignedDuration under its original name.

This type is not currently deprecated, but it likely will be in the future. Use SignedDuration when writing new code. In the next breaking release, this alias will be removed.

Aliased Type

#[repr(C)]
pub struct Duration {
    seconds: i64,
    nanoseconds: RangedI32<-999999999, 999999999>,
    _padding: Padding,
}

Fields
seconds: i64

Number of whole seconds.

nanoseconds: RangedI32<-999999999, 999999999>

Number of nanoseconds within the second. The sign always matches the seconds field.

_padding: Padding