time::unit

Struct Microsecond

pub struct Microsecond<const N: u128 = 1>;

A unit of time representing exactly N microseconds.

Implementations

impl Microsecond

pub const fn per<T>(_larger: T) -> <T as DefaultOutput<Microsecond>>::Output
where T: MultipleOf<Microsecond, <T as DefaultOutput<Microsecond>>::Output> + DefaultOutput<Microsecond> + Copy,

Obtain the number of times Microsecond can fit into T. If T is smaller than Microsecond, the code will fail to compile. The return type is the smallest unsigned integer type that can represent the value.

Valid calls:

Microsecond::per(Microsecond) (returns u8)

Microsecond::per(Millisecond) (returns u16)

Microsecond::per(Second) (returns u32)

Microsecond::per(Minute) (returns u32)

Microsecond::per(Hour) (returns u32)

Microsecond::per(Day) (returns u64)

Microsecond::per(Week) (returns u64)

pub const fn per_t<Output>( larger: impl MultipleOf<Microsecond, Output> + Copy, ) -> Output

Obtain the number of times Microsecond can fit into T. If T is smaller than Microsecond, the code will fail to compile. The return type is any primitive numeric type that can represent the value.

Valid calls:

Microsecond::per(Microsecond) (returns u8, u16, u32, u64, u128, usize, i8, i16, i32, i64, i128, isize, f32, or f64)

Microsecond::per(Millisecond) (returns u16, u32, u64, u128, usize, i16, i32, i64, i128, isize, f32, or f64)

Microsecond::per(Second) (returns u32, u64, u128, usize, i32, i64, i128, isize, f32, or f64)

Microsecond::per(Minute) (returns u32, u64, u128, usize, i32, i64, i128, isize, f32, or f64)

Microsecond::per(Hour) (returns u32, u64, u128, i64, i128, f32, or f64)

Microsecond::per(Day) (returns u64, u128, i64, i128, f32, or f64)

Microsecond::per(Week) (returns u64, u128, i64, i128, f32, or f64)