time::utc_offset

Struct UtcOffset

#[repr(C)]
pub struct UtcOffset {
    seconds: ri8<{ _ }, { _ }>,
    minutes: ri8<{ _ }, { _ }>,
    hours: ri8<-25, 25>,
}

An offset from UTC.

This struct can store values up to ±25:59:59. If you need support outside this range, please file an issue with your use case.

Fields
seconds: ri8<{ _ }, { _ }>minutes: ri8<{ _ }, { _ }>hours: ri8<-25, 25>
Implementations

impl UtcOffset

pub const UTC: Self

A UtcOffset that is UTC.

assert_eq!(UtcOffset::UTC, offset!(UTC));

pub(crate) const fn as_u32_for_equality(self) -> u32

Provide a representation of the UtcOffset as a i32. This value can be used for equality, and hashing. This value is not suitable for ordering; use as_i32_for_comparison instead.

const fn as_i32_for_comparison(self) -> i32

Provide a representation of the UtcOffset as a i32. This value can be used for ordering. While it is suitable for equality, as_u32_for_equality is preferred for performance reasons.

pub const fn from_hms( hours: i8, minutes: i8, seconds: i8, ) -> Result<Self, ComponentRange>

Create a UtcOffset representing an offset by the number of hours, minutes, and seconds provided.

The sign of all three components should match. If they do not, all smaller components will have their signs flipped.

assert_eq!(UtcOffset::from_hms(1, 2, 3)?.as_hms(), (1, 2, 3));
assert_eq!(UtcOffset::from_hms(1, -2, -3)?.as_hms(), (1, 2, 3));

pub(crate) const fn from_hms_ranged_unchecked( hours: ri8<-25, 25>, minutes: ri8<{ _ }, { _ }>, seconds: ri8<{ _ }, { _ }>, ) -> Self

Create a UtcOffset representing an offset of the hours, minutes, and seconds provided. All three parameters must have the same sign.

While the signs of the parameters are required to match, this is not a safety invariant.

pub(crate) const fn from_hms_ranged( hours: ri8<-25, 25>, minutes: ri8<{ _ }, { _ }>, seconds: ri8<{ _ }, { _ }>, ) -> Self

Create a UtcOffset representing an offset by the number of hours, minutes, and seconds provided.

The sign of all three components should match. If they do not, all smaller components will have their signs flipped.

pub const fn from_whole_seconds(seconds: i32) -> Result<Self, ComponentRange>

Create a UtcOffset representing an offset by the number of seconds provided.

assert_eq!(UtcOffset::from_whole_seconds(3_723)?.as_hms(), (1, 2, 3));

pub(crate) const fn from_whole_seconds_ranged( seconds: ri32<{ _ }, { _ }>, ) -> Self

Create a UtcOffset representing an offset by the number of seconds provided.

assert_eq!(
    UtcOffset::from_whole_seconds_ranged(RangedI32::new_static::<3_723>()).as_hms(),
    (1, 2, 3)
);

pub const fn as_hms(self) -> (i8, i8, i8)

Obtain the UTC offset as its hours, minutes, and seconds. The sign of all three components will always match. A positive value indicates an offset to the east; a negative to the west.

assert_eq!(offset!(+1:02:03).as_hms(), (1, 2, 3));
assert_eq!(offset!(-1:02:03).as_hms(), (-1, -2, -3));

pub(crate) const fn as_hms_ranged( self, ) -> (ri8<-25, 25>, ri8<{ _ }, { _ }>, ri8<{ _ }, { _ }>)

Obtain the UTC offset as its hours, minutes, and seconds. The sign of all three components will always match. A positive value indicates an offset to the east; a negative to the west.

pub const fn whole_hours(self) -> i8

Obtain the number of whole hours the offset is from UTC. A positive value indicates an offset to the east; a negative to the west.

assert_eq!(offset!(+1:02:03).whole_hours(), 1);
assert_eq!(offset!(-1:02:03).whole_hours(), -1);

pub const fn whole_minutes(self) -> i16

Obtain the number of whole minutes the offset is from UTC. A positive value indicates an offset to the east; a negative to the west.

assert_eq!(offset!(+1:02:03).whole_minutes(), 62);
assert_eq!(offset!(-1:02:03).whole_minutes(), -62);

pub const fn minutes_past_hour(self) -> i8

Obtain the number of minutes past the hour the offset is from UTC. A positive value indicates an offset to the east; a negative to the west.

assert_eq!(offset!(+1:02:03).minutes_past_hour(), 2);
assert_eq!(offset!(-1:02:03).minutes_past_hour(), -2);

pub const fn whole_seconds(self) -> i32

Obtain the number of whole seconds the offset is from UTC. A positive value indicates an offset to the east; a negative to the west.

assert_eq!(offset!(+1:02:03).whole_seconds(), 3723);
assert_eq!(offset!(-1:02:03).whole_seconds(), -3723);

pub const fn seconds_past_minute(self) -> i8

Obtain the number of seconds past the minute the offset is from UTC. A positive value indicates an offset to the east; a negative to the west.

assert_eq!(offset!(+1:02:03).seconds_past_minute(), 3);
assert_eq!(offset!(-1:02:03).seconds_past_minute(), -3);

pub const fn is_utc(self) -> bool

Check if the offset is exactly UTC.

assert!(!offset!(+1:02:03).is_utc());
assert!(!offset!(-1:02:03).is_utc());
assert!(offset!(UTC).is_utc());

pub const fn is_positive(self) -> bool

Check if the offset is positive, or east of UTC.

assert!(offset!(+1:02:03).is_positive());
assert!(!offset!(-1:02:03).is_positive());
assert!(!offset!(UTC).is_positive());

pub const fn is_negative(self) -> bool

Check if the offset is negative, or west of UTC.

assert!(!offset!(+1:02:03).is_negative());
assert!(offset!(-1:02:03).is_negative());
assert!(!offset!(UTC).is_negative());

pub fn local_offset_at( datetime: OffsetDateTime, ) -> Result<Self, IndeterminateOffset>

Attempt to obtain the system’s UTC offset at a known moment in time. If the offset cannot be determined, an error is returned.

let local_offset = UtcOffset::local_offset_at(OffsetDateTime::UNIX_EPOCH);
assert!(local_offset.is_ok());

pub fn current_local_offset() -> Result<Self, IndeterminateOffset>

Attempt to obtain the system’s current UTC offset. If the offset cannot be determined, an error is returned.

let local_offset = UtcOffset::current_local_offset();
assert!(local_offset.is_ok());

impl UtcOffset

pub fn format_into( self, output: &mut (impl Write + ?Sized), format: &(impl Formattable + ?Sized), ) -> Result<usize, Format>

Format the UtcOffset using the provided format description.

pub fn format( self, format: &(impl Formattable + ?Sized), ) -> Result<String, Format>

Format the UtcOffset using the provided format description.

let format =
    format_description::parse_borrowed::<3>("[offset_hour sign:mandatory]:[offset_minute]")?;
assert_eq!(offset!(+1).format(&format)?, "+01:00");

impl UtcOffset

pub fn parse( input: &str, description: &(impl Parsable + ?Sized), ) -> Result<Self, Parse>

Parse a UtcOffset from the input using the provided format description.

let format = format_description!("[offset_hour]:[offset_minute]");
assert_eq!(UtcOffset::parse("-03:42", &format)?, offset!(-3:42));

pub fn parse_with_defaults( input: &[u8], description: &(impl Parsable + ?Sized), defaults: Parsed, ) -> Result<Self, Parse>

Parse a UtcOffset from the input using the provided format description and default values.

let format = format_description!("[offset_hour sign:mandatory]");
let defaults = Parsed::new()
    .with_offset_minute_signed(30)
    .expect("30 is a valid offset minute");
assert_eq!(
    UtcOffset::parse_with_defaults(b"+05", &format, defaults)?,
    offset!(+5:30)
);

impl UtcOffset

pub(crate) const DISPLAY_BUFFER_SIZE: usize = 9

The maximum number of bytes that the fmt_into_buffer method will write, which is also used for the Display implementation.

pub(crate) const fn fmt_into_buffer( self, buf: &mut [MaybeUninit<u8>; 9], ) -> usize

Format the UtcOffset into the provided buffer, returning the number of bytes written.