Crate time

Feature flags

This crate exposes a number of features. These can be enabled or disabled as shown in Cargo’s documentation. Features are disabled by default unless otherwise noted.

Reliance on a given feature is always indicated alongside the item definition.

std (enabled by default, implicitly enables alloc)

This enables a number of features that depend on the standard library.

alloc (enabled by default via std)

Enables a number of features that require the ability to dynamically allocate memory.

macros

Enables macros that provide compile-time verification of values and intuitive syntax.

formatting (implicitly enables std)

Enables formatting of most structs.

parsing

Enables parsing of most structs.

local-offset (implicitly enables std)

This feature enables a number of methods that allow obtaining the system’s UTC offset.

large-dates

By default, only years within the ±9999 range (inclusive) are supported. If you need support for years outside this range, consider enabling this feature; the supported range will be increased to ±999,999.

Note that enabling this feature has some costs, as it means forgoing some optimizations. Ambiguities may be introduced when parsing that would not otherwise exist.

serde

Enables serde support for all types.

serde-human-readable (implicitly enables serde, formatting, and parsing)

Allows serde representations to use a human-readable format. This is determined by the serializer, not the user. If this feature is not enabled or if the serializer requests a non-human-readable format, a format optimized for binary representation will be used.

Libraries should never enable this feature, as the decision of what format to use should be up to the user.

rand (implicitly enables rand08 and rand09)

Previously, this would enable support for rand 0.8. Since the release of rand 0.9, the feature has been split into rand08 and rand09 to allow support for both versions. For backwards compatibility and simplicity, this feature enables support for both series.

It is strongly recommended to enable rand08 or rand09 directly, as enabling rand will needlessly pull in both versions.

rand08

Enables rand 0.8 support for all types.

rand09

Enables rand 0.9 support for all types.

quickcheck (implicitly enables alloc)

Enables quickcheck support for all types.

wasm-bindgen

Enables wasm-bindgen support for converting JavaScript dates, as well as obtaining the UTC offset from JavaScript.

Re-exports

pub use crate::error::Error;

Modules

convert

Conversion between units of time.

date 🔒

The Date struct and its associated impls.

duration 🔒

The Duration struct and its associated impls.

error

Various error types returned by methods in the time crate.

ext

Extension traits.

format_description

Description of how types should be formatted and parsed.

formatting

Formatting for various types.

hint 🔒

Hints to the compiler that affects how code should be emitted or optimized.

instant 🔒

The [Instant] struct and its associated impls.

internal_macros 🔒

Macros for use within the library. They are not publicly available.

interop 🔒

Comparison, arithmetic, and conversion between various types in time and the standard library.

macros

Macros to construct statically known values.

month 🔒

The Month enum and its associated impls.

offset_date_time 🔒

The OffsetDateTime struct and its associated impls.

parsing

Parsing for various types.

primitive_date_time 🔒

The PrimitiveDateTime struct and its associated impls.

serde

Differential formats for serde.

sys 🔒

Functions with a common interface that rely on system calls.

time 🔒

The Time struct and its associated impls.

utc_date_time 🔒

The UtcDateTime struct and associated impls.

utc_offset 🔒

The UtcOffset struct and its associated impls.

util

Utility functions, including updating time zone information.

weekday 🔒

Days of the week.

Structs

Date

Date in the proleptic Gregorian calendar.

Duration

A span of time with nanosecond precision.

OffsetDateTime

A PrimitiveDateTime with a UtcOffset.

PrimitiveDateTime

Combined date and time.

Time

The clock time within a given date. Nanosecond precision.

UtcDateTime

A PrimitiveDateTime that is known to be UTC.

UtcOffset

An offset from UTC.

Enums

Month

Months of the year.

Weekday

Days of the week.

Functions

panic 🔒

This is a separate function to reduce the code size of explicit panics.

size_of_val 🔒

Returns the size of the pointed-to value in bytes.

Type Aliases

Result

An alias for std::result::Result with a generic error from the time crate.