Crate time

This crate contains abstractions to interact with hardware clocks.

Structs

Duration

A Duration type to represent a span of time, typically used for system timeouts.

Frequency

A clock frequency.

Instant

A measurement of a monotonically nondecreasing clock.

Monotonic

WallTime

Traits

ClockSource

A clock source.

ClockType

Either a Monotonic or WallTime clock.

EarlySleeper

A hardware clock that can sleep without relying on interrupts.

Functions

early_sleep

Wait for the given duration.

now

Returns the current time.

register_clock_source

Register a clock source.

register_early_sleeper

Register the clock source that can be used to sleep when interrupts are disabled.