Module time::error

Various error types returned by methods in the time crate.

Structs

ComponentRange

An error type indicating that a component provided to a method was out of range, causing a failure.

ConversionRange

An error type indicating that a conversion failed because the target type could not store the initial value.

DifferentVariant

An error type indicating that a TryFrom call failed because the original value was of a different variant.

Enums

Error

A unified error type for anything returned by a method in the time crate.

Format

An error occurred when formatting.

InvalidFormatDescription

The format description provided was not valid.