Enum time::error::Parse

#[non_exhaustive]
pub enum Parse {
    TryFromParsed(TryFromParsed),
    ParseFromDescription(ParseFromDescription),
    UnexpectedTrailingCharacters,
}

An error that occurred at some stage of parsing.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

TryFromParsed(TryFromParsed)

ParseFromDescription(ParseFromDescription)

UnexpectedTrailingCharacters

The input should have ended, but there were characters remaining.