time
Module format_description

Available on crate features formatting or parsing only.

Description of how types should be formatted and parsed.

The formatted value will be output to the provided writer. Format descriptions can be well-known or obtained by using the format_description! macro or a function listed below.

For examples, see the implementors of Formattable, e.g. well_known::Rfc3339.

Modules

modifier

Various modifiers for components.

well_known

Well-known formats, typically standards.

Enums

BorrowedFormatItem

A complete description of how to format and parse a type.

Component

A component of a larger format description.

OwnedFormatItemalloc

A complete description of how to format and parse a type.

Functions

parsealloc

Parse a sequence of items from the format description.

parse_borrowedalloc

Parse a sequence of items from the format description.

parse_ownedalloc

Parse a sequence of items from the format description.

parse_strftime_borrowedalloc

Parse a sequence of items from the strftime format description.

parse_strftime_ownedalloc

Parse a sequence of items from the strftime format description.