Module time::format_description

Description of how types should be formatted and parsed.

The formatted value will be output to the provided writer. Format descriptions can be well-known or obtained by using the format_description! macro, the format_description::parse function.

Modules

modifier

Various modifiers for components.

well_known

Well-known formats, typically standards.

Enums

Component

A component of a larger format description.

FormatItem

A complete description of how to format and parse a type.

Functions

parse

Parse a sequence of items from the format description.