Enum time::format_description::Component

#[non_exhaustive]
pub enum Component {

    Day(Day),
    Month(Month),
    Ordinal(Ordinal),
    Weekday(Weekday),
    WeekNumber(WeekNumber),
    Year(Year),
    Hour(Hour),
    Minute(Minute),
    Period(Period),
    Second(Second),
    Subsecond(Subsecond),
    OffsetHour(OffsetHour),
    OffsetMinute(OffsetMinute),
    OffsetSecond(OffsetSecond),
    Ignore(Ignore),
    UnixTimestamp(UnixTimestamp),
    End(End),

}

A component of a larger format description.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Day(Day)

Day of the month.

Month(Month)

Month of the year.

Ordinal(Ordinal)

Ordinal day of the year.

Weekday(Weekday)

Day of the week.

WeekNumber(WeekNumber)

Week within the year.

Year(Year)

Year of the date.

Hour(Hour)

Hour of the day.

Minute(Minute)

Minute within the hour.

Period(Period)

AM/PM part of the time.

Second(Second)

Second within the minute.

Subsecond(Subsecond)

Subsecond within the second.

OffsetHour(OffsetHour)

Hour of the UTC offset.

OffsetMinute(OffsetMinute)

Minute within the hour of the UTC offset.

OffsetSecond(OffsetSecond)

Second within the minute of the UTC offset.

Ignore(Ignore)

A number of bytes to ignore when parsing. This has no effect on formatting.

UnixTimestamp(UnixTimestamp)

A Unix timestamp.

End(End)

The end of input. Parsing this component will fail if there is any input remaining. This component neither affects formatting nor consumes any input when parsing.