time::format_description

Function parse_borrowed

pub fn parse_borrowed<const VERSION: usize>(
    s: &str,
) -> Result<Vec<BorrowedFormatItem<'_>>, InvalidFormatDescription>

Available on crate feature alloc and (crate features formatting or parsing) only.

Parse a sequence of items from the format description.

The syntax for the format description can be found in the book. The version of the format description is provided as the const parameter. It is recommended to use version 2.