time::format_description

Module modifier

Available on crate features formatting or parsing only.

Various modifiers for components.

Structs

Day

Day of the month.

End

The end of input.

Hour

Hour of the day.

Ignore

Ignore some number of bytes.

Minute

Minute within the hour.

Month

Month of the year.

OffsetHour

Hour of the UTC offset.

OffsetMinute

Minute within the hour of the UTC offset.

OffsetSecond

Second within the minute of the UTC offset.

Ordinal

Ordinal day of the year.

Period

AM/PM part of the time.

Second

Second within the minute.

Subsecond

Subsecond within the second.

UnixTimestamp

A Unix timestamp.

WeekNumber

Week within the year.

Weekday

Day of the week.

Year

Year of the date.

Enums

MonthRepr

The representation of a month.

Padding

Type of padding to ensure a minimum width.

SubsecondDigits

The number of digits present in a subsecond representation.

UnixTimestampPrecision

The precision of a Unix timestamp.

WeekNumberRepr

The representation used for the week number.

WeekdayRepr

The representation used for the day of the week.

YearRepr

The representation used for a year value.