Module time::format_description::parse

Parser for format descriptions.

Modules

ast 🔒

AST for parsing format descriptions.

format_item 🔒

Typed, validated representation of a parsed format description.

lexer 🔒

Lexer for parsing format descriptions.

Macros

validate_version 🔒

A helper macro to statically validate the version (when used as a const parameter).

version 🔒

A helper macro to make version restrictions simpler to read and write.

Structs

Error 🔒

A complete error description.

ErrorInner 🔒

The internal error type.

Location 🔒

A location within a string.

Span 🔒

A start and end point within a string.

Spanned 🔒

A value with an associated Span.

Unused 🔒

A value that may be used in the future, but currently is not.

Version 🔒

A struct that is used to ensure that the version is valid.

Traits

SpannedValue 🔒

Helper trait to attach a Span to a value.

Functions

parse

Parse a sequence of items from the format description.

parse_borrowed

Parse a sequence of items from the format description.

parse_owned

Parse a sequence of items from the format description.

unused 🔒

Indicate that a value is currently unused.