time::format_description::well_known

Module iso8601

Available on crate features formatting or parsing only.

The format described in ISO 8601.

Structs

Config

Configuration for Iso8601.

Iso8601

The format described in ISO 8601.

Enums

DateKind

Which format to use for the date.

FormattedComponents

Which components to format.

OffsetPrecision

The precision for the UTC offset.

TimePrecision

The precision and number of decimal digits present for the time.

Type Aliases

EncodedConfig

An encoded Config that can be used as a const parameter to Iso8601.