Type Alias time::format_description::well_known::iso8601::EncodedConfig

pub type EncodedConfig = DoNotRelyOnWhatThisIs;

An encoded Config that can be used as a const parameter to Iso8601.

The type this is aliased to must not be relied upon. It can change in any release without notice.