time

Module formatting

Formatting for various types.

Modules

formattable ğŸ”’

A trait that can be used to format an item from its components.

iso8601 ğŸ”’

Helpers for implementing formatting for ISO 8601.

Constants

MONTH_NAMES ğŸ”’

WEEKDAY_NAMES ğŸ”’

Traits

Formattable

A type that describes a format.

Functions

f64_10_pow_x ğŸ”’

Helper function to obtain 10^x, guaranteeing determinism for x â‰¤ 9. For these cases, the function optimizes to a lookup table. For x â‰¥ 10, it falls back to 10_f64.powi(x). The only situation where this would occur is if the user explicitly requests such precision when configuring the ISO 8601 well known format. All other possibilities max out at nine digits.

fmt_day ğŸ”’

Format the day into the designated output.

fmt_hour ğŸ”’

Format the hour into the designated output.

fmt_minute ğŸ”’

Format the minute into the designated output.

fmt_month ğŸ”’

Format the month into the designated output.

fmt_offset_hour ğŸ”’

Format the offset hour into the designated output.

fmt_offset_minute ğŸ”’

Format the offset minute into the designated output.

fmt_offset_second ğŸ”’

Format the offset second into the designated output.

fmt_ordinal ğŸ”’

Format the ordinal into the designated output.

fmt_period ğŸ”’

Format the period into the designated output.

fmt_second ğŸ”’

Format the second into the designated output.

fmt_subsecond ğŸ”’

Format the subsecond into the designated output.

fmt_unix_timestamp ğŸ”’

Format the Unix timestamp into the designated output.

fmt_week_number ğŸ”’

Format the week number into the designated output.

fmt_weekday ğŸ”’

Format the weekday into the designated output.

fmt_year ğŸ”’

Format the year into the designated output.

format_component ğŸ”’

Format the provided component into the designated output. An Err will be returned if the component requires information that it does not provide or if the value cannot be output to the stream.

format_float ğŸ”’

Write the floating point number to the output, returning the number of bytes written.

format_number ğŸ”’

Format a number with the provided padding and width.

format_number_pad_none ğŸ”’

Format a number with no padding.

format_number_pad_space ğŸ”’

Format a number with the provided width and spaces as padding.

format_number_pad_zero ğŸ”’

Format a number with the provided width and zeros as padding.

write ğŸ”’

Write all bytes to the output, returning the number of bytes written.

write_if ğŸ”’

If pred is true, write all bytes to the output, returning the number of bytes written.

write_if_else ğŸ”’

If pred is true, write true_bytes to the output. Otherwise, write false_bytes.