Module time::formatting

This is supported on crate feature formatting only.

Formatting for various types.

Modules

formattable
A trait that can be used to format an item from its components.

Constants

MONTH_NAMES

WEEKDAY_NAMES

Traits

DigitCount
A trait that indicates the formatted width of the value can be determined.

Formattable
A type that can be formatted.

Functions

fmt_day
Format the day into the designated output.

fmt_hour
Format the hour into the designated output.

fmt_minute
Format the minute into the designated output.

fmt_month
Format the month into the designated output.

fmt_offset_hour
Format the offset hour into the designated output.

fmt_offset_minute
Format the offset minute into the designated output.

fmt_offset_second
Format the offset second into the designated output.

fmt_ordinal
Format the ordinal into the designated output.

fmt_period
Format the period into the designated output.

fmt_second
Format the second into the designated output.

fmt_subsecond
Format the subsecond into the designated output.

fmt_week_number
Format the week number into the designated output.

fmt_weekday
Format the weekday into the designated output.

fmt_year
Format the year into the designated output.

format_component
Format the provided component into the designated output. An Err will be returned if the component requires information that it does not provide or if the value cannot be output to the stream.

format_number
Format a number with the provided padding and width.

format_number_pad_space
Format a number with the provided width and spaces as padding.

format_number_pad_zero
Format a number with the provided width and zeros as padding.