time::formatting::formattable::sealed

Trait Sealed

pub trait Sealed {
    // Required method
    fn format_into(
        &self,
        output: &mut (impl Write + ?Sized),
        date: Option<Date>,
        time: Option<Time>,
        offset: Option<UtcOffset>,
    ) -> Result<usize, Format>;

    // Provided method
    fn format(
        &self,
        date: Option<Date>,
        time: Option<Time>,
        offset: Option<UtcOffset>,
    ) -> Result<String, Format> { ... }
}

Format the item using a format description, the intended output, and the various components.

Required Methods

fn format_into( &self, output: &mut (impl Write + ?Sized), date: Option<Date>, time: Option<Time>, offset: Option<UtcOffset>, ) -> Result<usize, Format>

Format the item into the provided output, returning the number of bytes written.

Provided Methods

fn format( &self, date: Option<Date>, time: Option<Time>, offset: Option<UtcOffset>, ) -> Result<String, Format>

Format the item directly to a String.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.