Trait time::formatting::formattable::sealed::Sealed

pub trait Sealed {
    // Required method
    fn format_into(
        &self,
        output: &mut impl Write,
        date: Option<Date>,
        time: Option<Time>,
        offset: Option<UtcOffset>
    ) -> Result<usize, Format>;

    // Provided method
    fn format(
        &self,
        date: Option<Date>,
        time: Option<Time>,
        offset: Option<UtcOffset>
    ) -> Result<String, Format> { ... }
}

Format the item using a format description, the intended output, and the various components.

Required Methods

fn format_into( &self, output: &mut impl Write, date: Option<Date>, time: Option<Time>, offset: Option<UtcOffset> ) -> Result<usize, Format>

Format the item into the provided output, returning the number of bytes written.

Provided Methods

fn format( &self, date: Option<Date>, time: Option<Time>, offset: Option<UtcOffset> ) -> Result<String, Format>

Format the item directly to a String.