Trait time::formatting::formattable::sealed::Sealed

pub trait Sealed {
    fn format_into(
        &self, 
        output: &mut impl Write, 
        date: Option<Date>, 
        time: Option<Time>, 
        offset: Option<UtcOffset>
    ) -> Result<usize, Format>;

    fn format(
        &self, 
        date: Option<Date>, 
        time: Option<Time>, 
        offset: Option<UtcOffset>
    ) -> Result<String, Format> { ... }
}

This is supported on crate feature formatting only.

Format the item using a format description, the intended output, and the various components.

Required methods

fn format_into(
&self,
output: &mut impl Write,
date: Option<Date>,
time: Option<Time>,
offset: Option<UtcOffset>
) -> Result<usize, Format>

Format the item into the provided output, returning the number of bytes written.

Provided methods

fn format(
&self,
date: Option<Date>,
time: Option<Time>,
offset: Option<UtcOffset>
) -> Result<String, Format>

Format the item directly to a String.