Module time::parsing::combinator

Implementations of the low-level parser combinators.

Modules

rfc 🔒

Combinators for rules as defined in a standard.

Functions

any_digit 🔒

Consume exactly one digit.

ascii_char 🔒

Consume exactly one of the provided ASCII characters.

ascii_char_ignore_case 🔒

Consume exactly one of the provided ASCII characters, case-insensitive.

exactly_n_digits 🔒

Consume exactly n digits, returning the numerical value.

exactly_n_digits_padded 🔒

Consume exactly n digits, returning the numerical value.

first_match 🔒

Consume the first matching item, returning its associated value.

n_to_m 🔒

Consume between n and m instances of the provided parser.

n_to_m_digits 🔒

Consume between n and m digits, returning the numerical value.

n_to_m_digits_padded 🔒

Consume between n and m digits, returning the numerical value.

one_or_more 🔒

Consume one of or more instances of the provided parser. The parser must produce the unit value.

opt 🔒

Optionally consume an input with a given parser.

sign 🔒

Parse a “+” or “-” sign. Returns the ASCII byte representing the sign, if present.

zero_or_more 🔒

Consume zero or more instances of the provided parser. The parser must return the unit value.