Trait time::parsing::parsable::sealed::Sealed

pub trait Sealed {
    fn parse_into<'a>(
        &self, 
        input: &'a [u8], 
        parsed: &mut Parsed
    ) -> Result<&'a [u8], Parse>;

    fn parse(&self, input: &[u8]) -> Result<Parsed, Parse> { ... }

    fn parse_date(&self, input: &[u8]) -> Result<Date, Parse> { ... }

    fn parse_time(&self, input: &[u8]) -> Result<Time, Parse> { ... }

    fn parse_offset(&self, input: &[u8]) -> Result<UtcOffset, Parse> { ... }

    fn parse_date_time(&self, input: &[u8]) -> Result<PrimitiveDateTime, Parse> { ... }

    fn parse_offset_date_time(
        &self, 
        input: &[u8]
    ) -> Result<OffsetDateTime, Parse> { ... }
}

This is supported on crate feature parsing only.

Parse the item using a format description and an input.

Required methods

fn parse_into<'a>(
&self,
input: &'a [u8],
parsed: &mut Parsed
) -> Result<&'a [u8], Parse>

Parse the item into the provided Parsed struct.

This method can be used to parse a single component without parsing the full value.

Provided methods

fn parse(&self, input: &[u8]) -> Result<Parsed, Parse>

Parse the item into a new Parsed struct.

This method can only be used to parse a complete value of a type. If any characters remain after parsing, an error will be returned.

fn parse_date(&self, input: &[u8]) -> Result<Date, Parse>

Parse a Date from the format description.

fn parse_time(&self, input: &[u8]) -> Result<Time, Parse>

Parse a Time from the format description.

fn parse_offset(&self, input: &[u8]) -> Result<UtcOffset, Parse>

Parse a UtcOffset from the format description.

fn parse_date_time(&self, input: &[u8]) -> Result<PrimitiveDateTime, Parse>

Parse a PrimitiveDateTime from the format description.

fn parse_offset_date_time(&self, input: &[u8]) -> Result<OffsetDateTime, Parse>

Parse a OffsetDateTime from the format description.