Struct time::parsing::Parsed

pub struct Parsed { /* private fields */ }

All information parsed.

This information is directly used to construct the final values.

Most users will not need think about this struct in any way. It is public to allow for manual control over values, in the instance that the default parser is insufficient.

Implementations

impl Parsed

pub const fn new() -> Self

Create a new instance of Parsed with no information known.

pub fn parse_item<'a>( &mut self, input: &'a [u8], item: &impl AnyFormatItem ) -> Result<&'a [u8], ParseFromDescription>

Parse a single FormatItem or OwnedFormatItem, mutating the struct. The remaining input is returned as the Ok value.

If a FormatItem::Optional or OwnedFormatItem::Optional is passed, parsing will not fail; the input will be returned as-is if the expected format is not present.

pub fn parse_items<'a>( &mut self, input: &'a [u8], items: &[impl AnyFormatItem] ) -> Result<&'a [u8], ParseFromDescription>

Parse a sequence of FormatItems or OwnedFormatItems, mutating the struct. The remaining input is returned as the Ok value.

This method will fail if any of the contained FormatItems or OwnedFormatItems fail to parse. self will not be mutated in this instance.

pub fn parse_literal<'a>( input: &'a [u8], literal: &[u8] ) -> Result<&'a [u8], ParseFromDescription>

Parse a literal byte sequence. The remaining input is returned as the Ok value.

pub fn parse_component<'a>( &mut self, input: &'a [u8], component: Component ) -> Result<&'a [u8], ParseFromDescription>

Parse a single component, mutating the struct. The remaining input is returned as the Ok value.

impl Parsed

Getter methods

pub const fn year(&self) -> Option<i32>

Obtain the named component.

pub const fn year_last_two(&self) -> Option<u8>

Obtain the named component.

pub const fn iso_year(&self) -> Option<i32>

Obtain the named component.

pub const fn iso_year_last_two(&self) -> Option<u8>

Obtain the named component.

pub const fn month(&self) -> Option<Month>

Obtain the named component.

pub const fn sunday_week_number(&self) -> Option<u8>

Obtain the named component.

pub const fn monday_week_number(&self) -> Option<u8>

Obtain the named component.

pub const fn iso_week_number(&self) -> Option<NonZeroU8>

Obtain the named component.

pub const fn weekday(&self) -> Option<Weekday>

Obtain the named component.

pub const fn ordinal(&self) -> Option<NonZeroU16>

Obtain the named component.

pub const fn day(&self) -> Option<NonZeroU8>

Obtain the named component.

pub const fn hour_24(&self) -> Option<u8>

Obtain the named component.

pub const fn hour_12(&self) -> Option<NonZeroU8>

Obtain the named component.

pub const fn hour_12_is_pm(&self) -> Option<bool>

Obtain the named component.

pub const fn minute(&self) -> Option<u8>

Obtain the named component.

pub const fn second(&self) -> Option<u8>

Obtain the named component.

pub const fn subsecond(&self) -> Option<u32>

Obtain the named component.

pub const fn offset_hour(&self) -> Option<i8>

Obtain the named component.

pub const fn unix_timestamp_nanos(&self) -> Option<i128>

Obtain the named component.

pub const fn offset_minute(&self) -> Option<u8>

👎Deprecated since 0.3.8: use parsed.offset_minute_signed() instead

Obtain the absolute value of the offset minute.

pub const fn offset_minute_signed(&self) -> Option<i8>

Obtain the offset minute as an i8.

pub const fn offset_second(&self) -> Option<u8>

👎Deprecated since 0.3.8: use parsed.offset_second_signed() instead

Obtain the absolute value of the offset second.

pub const fn offset_second_signed(&self) -> Option<i8>

Obtain the offset second as an i8.

impl Parsed

Setter methods

All setters return Option<()>, which is Some if the value was set, and None if not. The setters may fail if the value is invalid, though behavior is not guaranteed.

pub fn set_year(&mut self, value: i32) -> Option<()>

Set the named component.

pub fn set_year_last_two(&mut self, value: u8) -> Option<()>

Set the named component.

pub fn set_iso_year(&mut self, value: i32) -> Option<()>

Set the named component.

pub fn set_iso_year_last_two(&mut self, value: u8) -> Option<()>

Set the named component.

pub fn set_month(&mut self, value: Month) -> Option<()>

Set the named component.

pub fn set_sunday_week_number(&mut self, value: u8) -> Option<()>

Set the named component.

pub fn set_monday_week_number(&mut self, value: u8) -> Option<()>

Set the named component.

pub fn set_iso_week_number(&mut self, value: NonZeroU8) -> Option<()>

Set the named component.

pub fn set_weekday(&mut self, value: Weekday) -> Option<()>

Set the named component.

pub fn set_ordinal(&mut self, value: NonZeroU16) -> Option<()>

Set the named component.

pub fn set_day(&mut self, value: NonZeroU8) -> Option<()>

Set the named component.

pub fn set_hour_24(&mut self, value: u8) -> Option<()>

Set the named component.

pub fn set_hour_12(&mut self, value: NonZeroU8) -> Option<()>

Set the named component.

pub fn set_hour_12_is_pm(&mut self, value: bool) -> Option<()>

Set the named component.

pub fn set_minute(&mut self, value: u8) -> Option<()>

Set the named component.

pub fn set_second(&mut self, value: u8) -> Option<()>

Set the named component.

pub fn set_subsecond(&mut self, value: u32) -> Option<()>

Set the named component.

pub fn set_offset_hour(&mut self, value: i8) -> Option<()>

Set the named component.

pub fn set_unix_timestamp_nanos(&mut self, value: i128) -> Option<()>

Set the named component.

pub fn set_offset_minute(&mut self, value: u8) -> Option<()>

👎Deprecated since 0.3.8: use parsed.set_offset_minute_signed() instead

Set the named component.

pub fn set_offset_minute_signed(&mut self, value: i8) -> Option<()>

Set the offset_minute component.

pub fn set_offset_second(&mut self, value: u8) -> Option<()>

👎Deprecated since 0.3.8: use parsed.set_offset_second_signed() instead

Set the named component.

pub fn set_offset_second_signed(&mut self, value: i8) -> Option<()>

Set the offset_second component.

impl Parsed

Builder methods

All builder methods return Option<Self>, which is Some if the value was set, and None if not. The builder methods may fail if the value is invalid, though behavior is not guaranteed.

pub const fn with_year(self, value: i32) -> Option<Self>

Set the named component and return self.

pub const fn with_year_last_two(self, value: u8) -> Option<Self>

Set the named component and return self.

pub const fn with_iso_year(self, value: i32) -> Option<Self>

Set the named component and return self.

pub const fn with_iso_year_last_two(self, value: u8) -> Option<Self>

Set the named component and return self.

pub const fn with_month(self, value: Month) -> Option<Self>

Set the named component and return self.

pub const fn with_sunday_week_number(self, value: u8) -> Option<Self>

Set the named component and return self.

pub const fn with_monday_week_number(self, value: u8) -> Option<Self>

Set the named component and return self.

pub const fn with_iso_week_number(self, value: NonZeroU8) -> Option<Self>

Set the named component and return self.

pub const fn with_weekday(self, value: Weekday) -> Option<Self>

Set the named component and return self.

pub const fn with_ordinal(self, value: NonZeroU16) -> Option<Self>

Set the named component and return self.

pub const fn with_day(self, value: NonZeroU8) -> Option<Self>

Set the named component and return self.

pub const fn with_hour_24(self, value: u8) -> Option<Self>

Set the named component and return self.

pub const fn with_hour_12(self, value: NonZeroU8) -> Option<Self>

Set the named component and return self.

pub const fn with_hour_12_is_pm(self, value: bool) -> Option<Self>

Set the named component and return self.

pub const fn with_minute(self, value: u8) -> Option<Self>

Set the named component and return self.

pub const fn with_second(self, value: u8) -> Option<Self>

Set the named component and return self.

pub const fn with_subsecond(self, value: u32) -> Option<Self>

Set the named component and return self.

pub const fn with_offset_hour(self, value: i8) -> Option<Self>

Set the named component and return self.

pub const fn with_unix_timestamp_nanos(self, value: i128) -> Option<Self>

Set the named component and return self.

pub const fn with_offset_minute(self, value: u8) -> Option<Self>

👎Deprecated since 0.3.8: use parsed.with_offset_minute_signed() instead

Set the named component and return self.

pub const fn with_offset_minute_signed(self, value: i8) -> Option<Self>

Set the offset_minute component and return self.

pub const fn with_offset_second(self, value: u8) -> Option<Self>

👎Deprecated since 0.3.8: use parsed.with_offset_second_signed() instead

Set the named component and return self.

pub const fn with_offset_second_signed(self, value: i8) -> Option<Self>

Set the offset_second component and return self.