Module time::serde::timestamp::option

Treat an Option<OffsetDateTime> as a Unix timestamp for the purposes of serde.

Use this module in combination with serde’s #[with] attribute.

When deserializing, the offset is assumed to be UTC.

Functions

deserialize

Deserialize an Option<OffsetDateTime> from its Unix timestamp

serialize

Serialize an Option<OffsetDateTime> as its Unix timestamp