Struct time::Date

pub struct Date { /* fields omitted */ }

Calendar date.

Years between -100_000 and +100_000 inclusive are guaranteed to be representable. Any values outside this range may have incidental support that can change at any time without notice. If you need support outside this range, please file an issue with your use case.

Methods

impl Date

pub fn from_ymd(year: i32, month: u8, day: u8) -> Self

Create a Date from the year, month, and day.

assert_eq!(Date::from_ymd(2019, 1, 1), Date::from_yo(2019, 1));
assert_eq!(Date::from_ymd(2019, 12, 31), Date::from_yo(2019, 365));

Panics if the date is not valid.

Date::from_ymd(2019, 2, 29); // 2019 isn't a leap year.

pub fn from_yo(year: i32, ordinal: u16) -> Self

Create a Date from the year and ordinal day number.

assert_eq!(Date::from_yo(2019, 1), Date::from_ymd(2019, 1, 1));
assert_eq!(Date::from_yo(2019, 365), Date::from_ymd(2019, 12, 31));

Panics if the date is not valid.

Date::from_yo(2019, 366); // 2019 isn't a leap year.

pub fn from_iso_ywd(year: i32, week: u8, weekday: Weekday) -> Self

Create a Date from the ISO year, week, and weekday.

assert_eq!(Date::from_iso_ywd(2019, 1, Monday), Date::from_ymd(2018, 12, 31));
assert_eq!(Date::from_iso_ywd(2019, 1, Tuesday), Date::from_ymd(2019, 1, 1));
assert_eq!(Date::from_iso_ywd(2020, 53, Friday), Date::from_ymd(2021, 1, 1));

Panics if the week is not valid.

Date::from_iso_ywd(2019, 53, Monday); // 2019 doesn't have 53 weeks.

pub fn today() -> Self

Create a Date representing the current date.

assert!(Date::today().year() >= 2019);

pub fn year(self) -> i32

Get the year of the date.

assert_eq!(Date::from_ymd(2019, 1, 1).year(), 2019);
assert_eq!(Date::from_ymd(2019, 12, 31).year(), 2019);
assert_eq!(Date::from_ymd(2020, 1, 1).year(), 2020);

pub fn month(self) -> u8

Get the month of the date. If fetching both the month and day, use [Date::date] instead.

The returned value will always be in the range 1..=12.

assert_eq!(Date::from_ymd(2019, 1, 1).month(), 1);
assert_eq!(Date::from_ymd(2019, 12, 31).month(), 12);

pub fn day(self) -> u8

Get the day of the date. If fetching both the month and day, use [Date::date] instead.

The returned value will always be in the range 1..=31.

assert_eq!(Date::from_ymd(2019, 1, 1).day(), 1);
assert_eq!(Date::from_ymd(2019, 12, 31).day(), 31);

pub fn month_day(self) -> (u8, u8)

Get the month and day of the date.

The month component will always be in the range 1..=12; the day component in 1..=31.

assert_eq!(Date::from_ymd(2019, 1, 1).month_day(), (1, 1));
assert_eq!(Date::from_ymd(2019, 12, 31).month_day(), (12, 31));

pub fn ordinal(self) -> u16

Get the day of the year of the date.

The returned value will always be in the range 1..=366.

assert_eq!(Date::from_ymd(2019, 1, 1).ordinal(), 1);
assert_eq!(Date::from_ymd(2019, 12, 31).ordinal(), 365);

pub fn iso_year_week(self) -> (i32, u8)

Get the ISO 8601 year and week number.

assert_eq!(Date::from_ymd(2019, 1, 1).iso_year_week(), (2019, 1));
assert_eq!(Date::from_ymd(2019, 10, 4).iso_year_week(), (2019, 40));
assert_eq!(Date::from_ymd(2020, 1, 1).iso_year_week(), (2020, 1));
assert_eq!(Date::from_ymd(2020, 12, 31).iso_year_week(), (2020, 53));
assert_eq!(Date::from_ymd(2021, 1, 1).iso_year_week(), (2020, 53));

pub fn week(self) -> u8

Get the ISO week number of the date.

The returned value will always be in the range 1..=53.

assert_eq!(Date::from_ymd(2019, 1, 1).week(), 1);
assert_eq!(Date::from_ymd(2019, 10, 4).week(), 40);
assert_eq!(Date::from_ymd(2020, 1, 1).week(), 1);
assert_eq!(Date::from_ymd(2020, 12, 31).week(), 53);
assert_eq!(Date::from_ymd(2021, 1, 1).week(), 53);

pub fn sunday_based_week(self) -> u8

Get the week number where week 1 begins on the first Sunday.

The returned value will always be in the range 0..=53.

assert_eq!(Date::from_ymd(2019, 1, 1).sunday_based_week(), 0);
assert_eq!(Date::from_ymd(2020, 1, 1).sunday_based_week(), 0);
assert_eq!(Date::from_ymd(2020, 12, 31).sunday_based_week(), 52);
assert_eq!(Date::from_ymd(2021, 1, 1).sunday_based_week(), 0);

pub fn monday_based_week(self) -> u8

Get the week number where week 1 begins on the first Monday.

The returned value will always be in the range 0..=53.

assert_eq!(Date::from_ymd(2019, 1, 1).monday_based_week(), 0);
assert_eq!(Date::from_ymd(2020, 1, 1).monday_based_week(), 0);
assert_eq!(Date::from_ymd(2020, 12, 31).monday_based_week(), 52);
assert_eq!(Date::from_ymd(2021, 1, 1).monday_based_week(), 0);

pub fn as_ymd(self) -> (i32, u8, u8)

Get the year, month, and day of the date.

assert_eq!(Date::from_ymd(2019, 1, 1).as_ymd(), (2019, 1, 1));

pub fn as_yo(self) -> (i32, u16)

Get the year and ordinal day number of the date.

assert_eq!(Date::from_ymd(2019, 1, 1).as_yo(), (2019, 1));

pub fn weekday(self) -> Weekday

Get the weekday of the date.

This current uses Zeller's congruence internally.

assert_eq!(Date::from_ymd(2019, 1, 1).weekday(), Tuesday);
assert_eq!(Date::from_ymd(2019, 2, 1).weekday(), Friday);
assert_eq!(Date::from_ymd(2019, 3, 1).weekday(), Friday);
assert_eq!(Date::from_ymd(2019, 4, 1).weekday(), Monday);
assert_eq!(Date::from_ymd(2019, 5, 1).weekday(), Wednesday);
assert_eq!(Date::from_ymd(2019, 6, 1).weekday(), Saturday);
assert_eq!(Date::from_ymd(2019, 7, 1).weekday(), Monday);
assert_eq!(Date::from_ymd(2019, 8, 1).weekday(), Thursday);
assert_eq!(Date::from_ymd(2019, 9, 1).weekday(), Sunday);
assert_eq!(Date::from_ymd(2019, 10, 1).weekday(), Tuesday);
assert_eq!(Date::from_ymd(2019, 11, 1).weekday(), Friday);
assert_eq!(Date::from_ymd(2019, 12, 1).weekday(), Sunday);

pub fn next_day(self) -> Self

Get the next calendar date.

assert_eq!(Date::from_ymd(2019, 1, 1).next_day(), Date::from_ymd(2019, 1, 2));
assert_eq!(Date::from_ymd(2019, 1, 31).next_day(), Date::from_ymd(2019, 2, 1));
assert_eq!(Date::from_ymd(2019, 12, 31).next_day(), Date::from_ymd(2020, 1, 1));

pub fn previous_day(self) -> Self

Get the previous calendar date.

assert_eq!(Date::from_ymd(2019, 1, 2).previous_day(), Date::from_ymd(2019, 1, 1));
assert_eq!(Date::from_ymd(2019, 2, 1).previous_day(), Date::from_ymd(2019, 1, 31));
assert_eq!(Date::from_ymd(2020, 1, 1).previous_day(), Date::from_ymd(2019, 12, 31));

pub fn julian_day(self) -> i64

Get the Julian day for the date.

assert_eq!(Date::from_ymd(-4713, 11, 24).julian_day(), 0);
assert_eq!(Date::from_ymd(2000, 1, 1).julian_day(), 2_451_545);
assert_eq!(Date::from_ymd(2019, 1, 1).julian_day(), 2_458_485);
assert_eq!(Date::from_ymd(2019, 12, 31).julian_day(), 2_458_849);

pub fn from_julian_day(julian_day: i64) -> Self

Create a Date from the Julian day.

The algorithm to perform this conversion comes from E.G. Richards; it is available in section 15.11.3, courtesy of the United States Naval Observatory.

assert_eq!(Date::from_julian_day(0), Date::from_ymd(-4713, 11, 24));
assert_eq!(Date::from_julian_day(2_451_545), Date::from_ymd(2000, 1, 1));
assert_eq!(Date::from_julian_day(2_458_485), Date::from_ymd(2019, 1, 1));
assert_eq!(Date::from_julian_day(2_458_849), Date::from_ymd(2019, 12, 31));

impl Date

Methods to add a Time component, resulting in a DateTime.

pub const fn midnight(self) -> DateTime

Create a DateTime using the existing date. The Time component will be set to midnight.

assert_eq!(Date::from_ymd(1970, 1, 1).midnight(), DateTime::unix_epoch());

pub const fn with_time(self, time: Time) -> DateTime

Create a DateTime using the existing date and the provided Time.

assert_eq!(
    Date::from_ymd(1970, 1, 1).with_time(Time::from_hms(0, 0, 0)),
    Date::from_ymd(1970, 1, 1).midnight(),
);

pub fn with_hms(self, hour: u8, minute: u8, second: u8) -> DateTime

Create a DateTime using the existing date and the provided time.

assert_eq!(
    Date::from_ymd(1970, 1, 1).with_hms(0, 0, 0),
    Date::from_ymd(1970, 1, 1).with_time(Time::from_hms(0, 0, 0)),
);

pub fn with_hms_milli(
self,
hour: u8,
minute: u8,
second: u8,
millisecond: u16
) -> DateTime

Create a DateTime using the existing date and the provided time.

assert_eq!(
    Date::from_ymd(1970, 1, 1).with_hms_milli(0, 0, 0, 0),
    Date::from_ymd(1970, 1, 1).with_time(Time::from_hms_milli(0, 0, 0, 0)),
);

pub fn with_hms_micro(
self,
hour: u8,
minute: u8,
second: u8,
microsecond: u32
) -> DateTime

Create a DateTime using the existing date and the provided time.

assert_eq!(
    Date::from_ymd(1970, 1, 1).with_hms_micro(0, 0, 0, 0),
    Date::from_ymd(1970, 1, 1).with_time(Time::from_hms_micro(0, 0, 0, 0)),
);

pub fn with_hms_nano(
self,
hour: u8,
minute: u8,
second: u8,
nanosecond: u32
) -> DateTime

Create a DateTime using the existing date and the provided time.

assert_eq!(
    Date::from_ymd(1970, 1, 1).with_hms_nano(0, 0, 0, 0),
    Date::from_ymd(1970, 1, 1).with_time(Time::from_hms_nano(0, 0, 0, 0)),
);

impl Date

Methods that allow formatting the Date.

pub fn format(self, format: &str) -> String

Format the Date using the provided string. As no language is specified, English is used.

assert_eq!(Date::from_ymd(2019, 1, 2).format("%Y-%m-%d"), "2019-01-02");

pub fn format_language(self, format: &str, language: Language) -> String

Format the Date using the provided string and language.

assert_eq!(
    Date::from_ymd(2019, 1, 2).format_language("%B, %A", Language::en),
    "January, Wednesday",
);
assert_eq!(
    Date::from_ymd(2019, 1, 2).format_language("%B, %A", Language::es),
    "enero, miércoles",
);

pub fn parse(s: &str, format: &str) -> Result<Self, ParseError>

Attempt to parse a Date using the provided string.

assert_eq!(Date::parse("2019-01-02", "%F"), Ok(Date::from_ymd(2019, 1, 2)));
assert_eq!(Date::parse("2019-002", "%Y-%j"), Ok(Date::from_yo(2019, 2)));
assert_eq!(Date::parse("2019-W01-3", "%G-W%V-%u"), Ok(Date::from_iso_ywd(2019, 1, Wednesday)));

pub fn parse_language(
s: &str,
format: &str,
language: Language
) -> Result<Self, ParseError>

Attempt to parse a Date using the provided string.

assert_eq!(Date::parse_language("January 2 2019", "%B %-d %Y", en), Ok(Date::from_ymd(2019, 1, 2)));
assert_eq!(Date::parse_language("2 de enero 2019", "%-d de %B %Y", es), Ok(Date::from_ymd(2019, 1, 2)));