Struct time::Instant

pub struct Instant { /* private fields */ }

A measurement of a monotonically nondecreasing clock.

The inner value usually represents the internal counter value but the type is intentionally opaque and only useful with Duration.

Implementations

impl Instant

pub const ZERO: Self = _

pub const MAX: Self = _

pub fn new(counter: u64) -> Self

pub fn duration_since(&self, earlier: Self) -> Duration

Returns the amount of time elapsed from another instant to this one, or zero duration if that instant is later than this one.

pub fn checked_duration_since(&self, earlier: Self) -> Option<Duration>