Struct time::Time

pub struct Time { /* fields omitted */ }

The clock time within a given date. Nanosecond precision.

All minutes are assumed to have exactly 60 seconds; no attempt is made to handle leap seconds (either positive or negative).

As order is dependent on context (is noon before or after midnight?), this type does not implement PartialOrd or Ord.

Methods

impl Time

pub const fn midnight() -> Self

Create a Time that is exactly midnight.

assert_eq!(Time::midnight(), Time::from_hms(0, 0, 0));

pub fn from_hms(hour: u8, minute: u8, second: u8) -> Self

Create a Time from the hour, minute, and second.

let time = Time::from_hms(1, 2, 3);
assert_eq!(time.hour(), 1);
assert_eq!(time.minute(), 2);
assert_eq!(time.second(), 3);
assert_eq!(time.nanosecond(), 0);

Panics if any component is not valid.

Time::from_hms(24, 0, 0); // 24 isn't a valid hour.

Time::from_hms(0, 60, 0); // 60 isn't a valid minute.

Time::from_hms(0, 0, 60); // 60 isn't a valid second.

pub fn from_hms_milli(
hour: u8,
minute: u8,
second: u8,
millisecond: u16
) -> Self

Create a Time from the hour, minute, second, and millisecond.

let time = Time::from_hms_milli(1, 2, 3, 4);
assert_eq!(time.hour(), 1);
assert_eq!(time.minute(), 2);
assert_eq!(time.second(), 3);
assert_eq!(time.millisecond(), 4);
assert_eq!(time.nanosecond(), 4_000_000);

Panics if any component is not valid.

Time::from_hms_milli(24, 0, 0, 0); // 24 isn't a valid hour.

Time::from_hms_milli(0, 60, 0, 0); // 60 isn't a valid minute.

Time::from_hms_milli(0, 0, 60, 0); // 60 isn't a valid second.

Time::from_hms_milli(0, 0, 0, 1_000); // 1_000 isn't a valid millisecond.

pub fn from_hms_micro(
hour: u8,
minute: u8,
second: u8,
microsecond: u32
) -> Self

Create a Time from the hour, minute, second, and microsecond.

let time = Time::from_hms_micro(1, 2, 3, 4);
assert_eq!(time.hour(), 1);
assert_eq!(time.minute(), 2);
assert_eq!(time.second(), 3);
assert_eq!(time.microsecond(), 4);
assert_eq!(time.nanosecond(), 4_000);

Panics if any component is not valid.

Time::from_hms_micro(24, 0, 0, 0); // 24 isn't a valid hour.

Time::from_hms_micro(0, 60, 0, 0); // 60 isn't a valid minute.

Time::from_hms_micro(0, 0, 60, 0); // 60 isn't a valid second.

Time::from_hms_micro(0, 0, 0, 1_000_000); // 1_000_000 isn't a valid microsecond.

pub fn from_hms_nano(hour: u8, minute: u8, second: u8, nanosecond: u32) -> Self

Create a Time from the hour, minute, second, and nanosecond.

let time = Time::from_hms_nano(1, 2, 3, 4);
assert_eq!(time.hour(), 1);
assert_eq!(time.minute(), 2);
assert_eq!(time.second(), 3);
assert_eq!(time.nanosecond(), 4);

Panics if any component is not valid.

Time::from_hms_nano(24, 0, 0, 0); // 24 isn't a valid hour.

Time::from_hms_nano(0, 60, 0, 0); // 60 isn't a valid minute.

Time::from_hms_nano(0, 0, 60, 0); // 60 isn't a valid second.

Time::from_hms_nano(0, 0, 0, 1_000_000_000); // 1_000_000_000 isn't a valid nanosecond.

pub fn now() -> Self

Create a Time representing the current time (UTC).

println!("{:?}", Time::now());

pub const fn hour(self) -> u8

Returns the clock hour.

The returned value will always be in the range 0..=23.

assert_eq!(Time::from_hms(0, 0, 0).hour(), 0);
assert_eq!(Time::from_hms(23, 59, 59).hour(), 23);

pub const fn minute(self) -> u8

Returns the minute within the hour.

The returned value will always be in the range 0..60.

assert_eq!(Time::from_hms(0, 0, 0).minute(), 0);
assert_eq!(Time::from_hms(23, 59, 59).minute(), 59);

pub const fn second(self) -> u8

Returns the second within the minute.

The returned value will always be in the range 0..60.

assert_eq!(Time::from_hms(0, 0, 0).second(), 0);
assert_eq!(Time::from_hms(23, 59, 59).second(), 59);

pub const fn millisecond(self) -> u16

Return the milliseconds within the second.

The returned value will always be in the range 0..1_000.

assert_eq!(Time::from_hms_milli(0, 0, 0, 0).millisecond(), 0);
assert_eq!(Time::from_hms_milli(23, 59, 59, 999).millisecond(), 999);

pub const fn microsecond(self) -> u32

Return the microseconds within the second.

The returned value will always be in the range 0..1_000_000.

assert_eq!(Time::from_hms_micro(0, 0, 0, 0).microsecond(), 0);
assert_eq!(Time::from_hms_micro(23, 59, 59, 999_999).microsecond(), 999_999);

pub const fn nanosecond(self) -> u32

Return the nanoseconds within the second.

The returned value will always be in the range 0..1_000_000_000.

assert_eq!(Time::from_hms_nano(0, 0, 0, 0).nanosecond(), 0);
assert_eq!(Time::from_hms_nano(23, 59, 59, 999_999_999).nanosecond(), 999_999_999);

impl Time

Methods that allow formatting the Time.

pub fn format(self, format: &str) -> String

Format the Time using the provided string.

assert_eq!(Time::from_hms(0, 0, 0).format("%r"), "12:00:00 am");

pub fn parse(s: &str, format: &str) -> Result<Self, ParseError>

Attempt to parse a Time using the provided string.

assert_eq!(Time::parse("0:00:00", "%T"), Ok(Time::from_hms(0, 0, 0)));
assert_eq!(Time::parse("23:59:59", "%T"), Ok(Time::from_hms(23, 59, 59)));
assert_eq!(Time::parse("12:00:00 am", "%r"), Ok(Time::from_hms(0, 0, 0)));
assert_eq!(Time::parse("12:00:00 pm", "%r"), Ok(Time::from_hms(12, 0, 0)));
assert_eq!(Time::parse("11:59:59 pm", "%r"), Ok(Time::from_hms(23, 59, 59)));