Struct time::Tm

#[repr(C)]
pub struct Tm {
    pub tm_sec: i32,
    pub tm_min: i32,
    pub tm_hour: i32,
    pub tm_mday: i32,
    pub tm_mon: i32,
    pub tm_year: i32,
    pub tm_wday: i32,
    pub tm_yday: i32,
    pub tm_isdst: i32,
    pub tm_utcoff: i32,
    pub tm_nsec: i32,
}

Holds a calendar date and time broken down into its components (year, month, day, and so on), also called a broken-down time value.

Fields
tm_sec: i32

Seconds after the minute - [0, 60]

tm_min: i32

Minutes after the hour - [0, 59]

tm_hour: i32

Hours after midnight - [0, 23]

tm_mday: i32

Day of the month - [1, 31]

tm_mon: i32

Months since January - [0, 11]

tm_year: i32

Years since 1900

tm_wday: i32

Days since Sunday - [0, 6]. 0 = Sunday, 1 = Monday, …, 6 = Saturday.

tm_yday: i32

Days since January 1 - [0, 365]

tm_isdst: i32

Daylight Saving Time flag.

This value is positive if Daylight Saving Time is in effect, zero if Daylight Saving Time is not in effect, and negative if this information is not available.

tm_utcoff: i32

Identifies the time zone that was used to compute this broken-down time value, including any adjustment for Daylight Saving Time. This is the number of seconds east of UTC. For example, for U.S. Pacific Daylight Time, the value is -7*60*60 = -25200.

tm_nsec: i32

Nanoseconds after the second - [0, 109 - 1]

Implementations

impl Tm

pub fn to_timespec(&self) -> Timespec

Convert time to the seconds from January 1, 1970

pub fn to_local(&self) -> Tm

Convert time to the local timezone

pub fn to_utc(&self) -> Tm

Convert time to the UTC

pub fn ctime(&self) -> TmFmt<'_>

Returns a TmFmt that outputs according to the asctime format in ISO C, in the local timezone.

Example: “Thu Jan 1 00:00:00 1970”

pub fn asctime(&self) -> TmFmt<'_>

Returns a TmFmt that outputs according to the asctime format in ISO C.

Example: “Thu Jan 1 00:00:00 1970”

pub fn strftime<'a>(&'a self, format: &'a str) -> Result<TmFmt<'a>, ParseError>

Formats the time according to the format string.

pub fn rfc822(&self) -> TmFmt<'_>

Returns a TmFmt that outputs according to RFC 822.

local: “Thu, 22 Mar 2012 07:53:18 PST” utc: “Thu, 22 Mar 2012 14:53:18 GMT”

pub fn rfc822z(&self) -> TmFmt<'_>

Returns a TmFmt that outputs according to RFC 822 with Zulu time.

local: “Thu, 22 Mar 2012 07:53:18 -0700” utc: “Thu, 22 Mar 2012 14:53:18 -0000”

pub fn rfc3339<'a>(&'a self) -> TmFmt<'_>

Returns a TmFmt that outputs according to RFC 3339. RFC 3339 is compatible with ISO 8601.

local: “2012-02-22T07:53:18-07:00” utc: “2012-02-22T14:53:18Z”