Module time::util

Utility functions.

Functions

days_in_year
Get the number of calendar days in a given year.

is_leap_year
Returns if the provided year is a leap year in the proleptic Gregorian calendar. Uses astronomical year numbering.

validate_format_string
Checks if a user-provided formatting string is valid. If it isn't, a description of the error is returned.

weeks_in_year
Get the number of weeks in the ISO year.